Added new elements and changed site flow
