+++ docs/index.html
@@ -1,2 +1,1 @@
-<html> <head> <meta charSet="utf-8"> <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=0, maximum-scale=1, minimum-scale=1"> <meta name="description" content="Tubular React"> <meta name="keywords" content="application, html5, reactJS, responsive, web api"> <meta name="author" content="Unosquare Labs"> <link rel="shortcut icon" href="/tubular-react/favicon.2af2dfd9.ico"> <title>Tubular React | Unosquare Open Source</title> <style>#app,body,html{height:100%;font-family:Roboto}body{margin:0}</style> <link rel="stylesheet" href="/tubular-react/pages.9f05947c.css"></head> <body> <div id="app"></div> <script src="/tubular-react/pages.0609791e.js" type="text/javascript"></script> </body> </html>
-
+<html> <head> <meta charSet="utf-8"> <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=0, maximum-scale=1, minimum-scale=1"> <meta name="description" content="Tubular React"> <meta name="keywords" content="application, html5, reactJS, responsive, web api"> <meta name="author" content="Unosquare Labs"> <link rel="shortcut icon" href="/tubular-react/favicon.2af2dfd9.ico"> <title>Tubular React | Unosquare Open Source</title> <style>#app,body,html{height:100%;font-family:Roboto}body{margin:0}</style> <link rel="stylesheet" href="/tubular-react/pages.9f05947c.css"></head> <body> <div id="app"></div> <script src="/tubular-react/pages.08d25df3.js" type="text/javascript"></script> </body> <script>!function(e,a,n,t,c,g,o){e.GoogleAnalyticsObject=c,e.ga=e.ga||function(){(e.ga.q=e.ga.q||[]).push(arguments)},e.ga.l=1*new Date,g=a.createElement(n),o=a.getElementsByTagName(n)[0],g.async=1,g.src="//www.google-analytics.com/analytics.js",o.parentNode.insertBefore(g,o)}(window,document,"script",0,"ga"),ga("create","UA-8535255-1","unosquare.com"),ga("send","pageview");</script> </html>
