+++ docs/pages.08d25df3.js
@@ -942,25 +942,13 @@ var e=arguments[3];Object.defineProperty(exports,"__esModule",{value:!0}),export
 "use strict";var e=require("@babel/runtime/helpers/interopRequireDefault");Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var r=e(require("react")),t=e(require("recompose/pure")),u=e(require("@material-ui/core/SvgIcon"));function a(e,a){var i=function(t){return r.default.createElement(u.default,t,e)};return i.displayName="".concat(a,"Icon"),(i=(0,t.default)(i)).muiName="SvgIcon",i}var i=a;exports.default=i;
 },{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","recompose/pure":"Q3lV","@material-ui/core/SvgIcon":"qv6+"}],"KAKw":[function(require,module,exports) {
 "use strict";var e=require("@babel/runtime/helpers/interopRequireDefault");Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var t=e(require("react")),r=e(require("./utils/createSvgIcon")),l=(0,r.default)(t.default.createElement(t.default.Fragment,null,t.default.createElement("path",{d:"M15.41 7.41L14 6l-6 6 6 6 1.41-1.41L10.83 12z"}),t.default.createElement("path",{fill:"none",d:"M0 0h24v24H0z"})),"ChevronLeft");exports.default=l;
-},{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","./utils/createSvgIcon":"Atth"}],"UXoc":[function(require,module,exports) {
-"use strict";var e=require("@babel/runtime/helpers/interopRequireDefault");Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var t=e(require("react")),r=e(require("./utils/createSvgIcon")),l=(0,r.default)(t.default.createElement(t.default.Fragment,null,t.default.createElement("path",{d:"M12 8l-6 6 1.41 1.41L12 10.83l4.59 4.58L18 14z"}),t.default.createElement("path",{fill:"none",d:"M0 0h24v24H0z"})),"ExpandLess");exports.default=l;
-},{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","./utils/createSvgIcon":"Atth"}],"/evf":[function(require,module,exports) {
-"use strict";var e=require("@babel/runtime/helpers/interopRequireDefault");Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var t=e(require("react")),r=e(require("./utils/createSvgIcon")),a=(0,r.default)(t.default.createElement(t.default.Fragment,null,t.default.createElement("path",{d:"M16.59 8.59L12 13.17 7.41 8.59 6 10l6 6 6-6z"}),t.default.createElement("path",{fill:"none",d:"M0 0h24v24H0z"})),"ExpandMore");exports.default=a;
 },{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","./utils/createSvgIcon":"Atth"}],"O5EH":[function(require,module,exports) {
 "use strict";var e=require("@babel/runtime/helpers/interopRequireDefault");Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var t=e(require("react")),r=e(require("./utils/createSvgIcon")),a=(0,r.default)(t.default.createElement(t.default.Fragment,null,t.default.createElement("path",{fill:"none",d:"M0 0h24v24H0z"}),t.default.createElement("path",{d:"M3 18h18v-2H3v2zm0-5h18v-2H3v2zm0-7v2h18V6H3z"})),"Menu");exports.default=a;
 },{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","./utils/createSvgIcon":"Atth"}],"qM49":[function(require,module,exports) {
 "use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var r={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(r[t]=e[t]);return r.default=e,r};exports.__esModule=!0;var r=require("@material-ui/core"),t=e(require("react"));exports.default=function(e){return t.createElement(r.SvgIcon,e,t.createElement("path",{d:"M12 .3a12 12 0 0 0-3.8 23.4c.6.1.8-.3.8-.6v-2c-3.3.7-4-1.6-4-1.6-.6-1.4-1.4-1.8-1.4-1.8-1-.7.1-.7.1-.7 1.2 0 1.9 1.2 1.9 1.2 1 1.8 2.8 1.3 3.5 1 0-.8.4-1.3.7-1.6-2.7-.3-5.5-1.3-5.5-6 0-1.2.5-2.3 1.3-3.1-.2-.4-.6-1.6 0-3.2 0 0 1-.3 3.4 1.2a11.5 11.5 0 0 1 6 0c2.3-1.5 3.3-1.2 3.3-1.2.6 1.6.2 2.8 0 3.2.9.8 1.3 1.9 1.3 3.2 0 4.6-2.8 5.6-5.5 5.9.5.4.9 1 .9 2.2v3.3c0 .3.1.7.8.6A12 12 0 0 0 12 .3"}))};
 },{"@material-ui/core":"J2ra","react":"1n8/"}],"NH2g":[function(require,module,exports) {
-"use strict";var e=this&&this.__extends||function(){var e=function(t,r){return(e=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var r in t)t.hasOwnProperty(r)&&(e[r]=t[r])})(t,r)};return function(t,r){function n(){this.constructor=t}e(t,r),t.prototype=null===r?Object.create(r):(n.prototype=r.prototype,new n)}}(),t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}},r=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var r in e)Object.hasOwnProperty.call(e,r)&&(t[r]=e[r]);return t.default=e,t};exports.__esModule=!0;var n=require("@material-ui/core"),a=t(require("@material-ui/core/styles/createStyles")),i=t(require("@material-ui/icons/ChevronLeft")),l=t(require("@material-ui/icons/ExpandLess")),o=t(require("@material-ui/icons/ExpandMore")),s=t(require("@material-ui/icons/Menu")),c=r(require("react")),u=require("react-router-dom"),m=t(require("./Github")),p=function(){return a.default({flex:{flex:1,textAlign:"center"},image:{maxWidth:100},link:{textDecoration:"none"},logo:{maxWidth:150},nested:{marginLeft:10},spacer:{flex:"1 1 100%"}})},E=n.withStyles(p)(function(t){function r(){var e=null!==t&&t.apply(this,arguments)||this;return e.state={openDrawer:!1,openList:!1},e.toggleDrawer=function(t){return function(){e.setState({openDrawer:t}),e.state.openList&&e.setState({openList:t})}},e.toggleList=function(t){return function(){e.setState({openList:t})}},e}return e(r,t),r.prototype.render=function(){var e=this.props.classes;return c.createElement("div",null,c.createElement(n.AppBar,{position:"static",color:"primary"},c.createElement(n.Toolbar,null,c.createElement(n.Tooltip,{title:"Open"},c.createElement(n.IconButton,{onClick:this.toggleDrawer(!0),color:"inherit"},c.createElement(s.default,null))),c.createElement("img",{className:e.logo,src:"https://unosquare.github.io/tubular-react/static/tubular.png",alt:"Tubular"}),c.createElement("div",{className:e.spacer}),c.createElement(n.Tooltip,{title:"Tubular GitHub Repo"},c.createElement(n.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",className:e.flex,color:"inherit"},c.createElement(m.default,null))))),c.createElement(n.Drawer,{variant:"persistent",open:this.state.openDrawer,onClose:this.toggleDrawer(!1)},c.createElement(n.List,null,c.createElement(n.Tooltip,{title:"Close"},c.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},c.createElement(n.ListItemIcon,null,c.createElement(i.default,null)))),c.createElement(n.Divider,null),c.createElement(u.Link,{to:"/",className:e.link},c.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},c.createElement(n.ListItemText,{primary:"Home"}))),c.createElement(u.Link,{to:"/sample",className:e.link},c.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},c.createElement(n.ListItemText,{primary:"Samples"}))),c.createElement(n.ListItem,{button:!0,onClick:this.toggleList(!this.state.openList)},c.createElement(n.ListItemText,{primary:"Documentation"}),this.state.openList?c.createElement(l.default,null):c.createElement(o.default,null)),c.createElement(n.Collapse,{in:this.state.openList,timeout:"auto",unmountOnExit:!0},c.createElement(n.List,{component:"div",disablePadding:!0},c.createElement(u.Link,{to:"/documentation/getting-started",className:e.link},c.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},c.createElement(n.ListItemText,{primary:"Getting Started"}))),c.createElement(u.Link,{to:"/documentation/columnmodel",className:e.link},c.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},c.createElement(n.ListItemText,{primary:"Column Model"}))),c.createElement(u.Link,{to:"/documentation/datasource",className:e.link},c.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},c.createElement(n.ListItemText,{primary:"Data Source"}))),c.createElement(u.Link,{to:"/documentation/props",className:e.link},c.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},c.createElement(n.ListItemText,{primary:"Props"}))))),c.createElement(n.Divider,null),c.createElement(n.ListItem,{button:!0,component:"a",href:"https://www.unosquare.com/",target:"_blank",onClick:this.toggleDrawer(!1)},c.createElement("img",{className:e.image,src:"https://unosquare.github.io/tubular-react/static/logoUnosquare.png",alt:"Unosquare"})))))},r}(c.Component));exports.default=n.withStyles(p)(E);
-},{"@material-ui/core":"J2ra","@material-ui/core/styles/createStyles":"zEWA","@material-ui/icons/ChevronLeft":"KAKw","@material-ui/icons/ExpandLess":"UXoc","@material-ui/icons/ExpandMore":"/evf","@material-ui/icons/Menu":"O5EH","react":"1n8/","react-router-dom":"obMO","./Github":"qM49"}],"AkVR":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var l={};if(null!=e)for(var a in e)Object.hasOwnProperty.call(e,a)&&(l[a]=e[a]);return l.default=e,l};exports.__esModule=!0;var l=require("@material-ui/core"),a=e(require("react")),t=function(e){return l.createStyles({code:{background:"#F8F8FF",fontSize:14},root:{marginTop:3*e.spacing.unit,overflowX:"auto",width:"100%"}})},c=function(e){var t=e.classes;return a.createElement(l.Table,null,a.createElement(l.TableHead,null,a.createElement(l.TableRow,null,a.createElement(l.TableCell,null,"Name"),a.createElement(l.TableCell,null,"Type"),a.createElement(l.TableCell,null,"Default"),a.createElement(l.TableCell,null,"Description"),a.createElement(l.TableCell,null,"Options"))),a.createElement(l.TableBody,null,a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Name")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"string")),a.createElement(l.TableCell,null,"NONE"),a.createElement(l.TableCell,null,"This is required and represents a property of the entity which we are binding"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Aggregate")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"AggregateFunctions")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"NONE")),a.createElement(l.TableCell,null,"The aggregation function that will be applied to this column"),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"NONE"),",",a.createElement("code",{className:t.code},"SUM"),",",a.createElement("code",{className:t.code},"AVERAGE"),",",a.createElement("br",null),a.createElement("code",{className:t.code},"COUNT"),",",a.createElement("code",{className:t.code},"DISTINCT_COUNT"),",",a.createElement("br",null),a.createElement("code",{className:t.code},"MAX"),",",a.createElement("code",{className:t.code},"MIN"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"DataType")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"ColumnDataType")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"string")),a.createElement(l.TableCell,null,"The column type"),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"STRING"),",",a.createElement("code",{className:t.code},"NUMERIC"),",",a.createElement("code",{className:t.code},"BOOLEAN"),",",a.createElement("br",null),a.createElement("code",{className:t.code},"DATE"),",",a.createElement("code",{className:t.code},"DATE_TIME"),",",a.createElement("code",{className:t.code},"DATE_TIME_UTC"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Filterable")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"false")),a.createElement(l.TableCell,null,"Enables Filterable"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"IsKey")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"false")),a.createElement(l.TableCell,null,"Defines if a column is an identifier or not"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Label")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"string")),a.createElement(l.TableCell,null,"The name of the column"),a.createElement(l.TableCell,null,"Column label that will be shown"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Searchable")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,"Indicates that a column can be used to search upon"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"SortDirection")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"ColumnSortDirection")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"NONE")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"NONE"),",",a.createElement("code",{className:t.code},"ASCENDING"),",",a.createElement("code",{className:t.code},"DESCENDING"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"SortOrder")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"number")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"-1")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Sortable")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"false")),a.createElement(l.TableCell,null,"Determines if a column can be sorted"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Visible")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,"Specifies if a column should be shown"),a.createElement(l.TableCell,null,"-"))))};exports.default=l.withStyles(t)(c);
-},{"@material-ui/core":"J2ra","react":"1n8/"}],"5NpU":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var r in e)Object.hasOwnProperty.call(e,r)&&(t[r]=e[r]);return t.default=e,t};exports.__esModule=!0;var t=require("@material-ui/core"),r=require("@material-ui/core/styles"),a=e(require("react")),n=require("react-router-dom"),i={container:{marginTop:52,padding:30},link:{textDecoration:"none"},paper:{padding:10}},l=function(e){var r=e.classes;return a.createElement(t.Paper,{className:r.paper},a.createElement(t.List,null,a.createElement(t.Typography,{variant:"h5"},"Documentation"),a.createElement(t.Divider,null),a.createElement(n.Link,{to:"/Documentation/getting-started",className:r.link},a.createElement(t.ListItem,{button:!0},a.createElement(t.ListItemText,{primary:"Getting Started"}))),a.createElement(n.Link,{to:"/Documentation/datasource",className:r.link},a.createElement(t.ListItem,{button:!0},a.createElement(t.ListItemText,{primary:"Data Source"}))),a.createElement(n.Link,{to:"/Documentation/props",className:r.link},a.createElement(t.ListItem,{button:!0},a.createElement(t.ListItemText,{primary:"Props"}))),a.createElement(n.Link,{to:"/Documentation/columnmodel",className:r.link},a.createElement(t.ListItem,{button:!0},a.createElement(t.ListItemText,{primary:"Column Model"})))))};exports.default=r.withStyles(i)(l);
-},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/","react-router-dom":"obMO"}],"q1fd":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var r in e)Object.hasOwnProperty.call(e,r)&&(t[r]=e[r]);return t.default=e,t},t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var r=require("@material-ui/core"),a=require("@material-ui/core/styles"),n=e(require("react")),l=t(require("../../components/ColumnModelTable")),i=t(require("../../components/DocumentationList")),u={container:{margin:"0",padding:30,width:"100%"},paper:{padding:10}},o=function(e){var t=e.classes;return n.createElement("div",null,n.createElement(r.Grid,{container:!0,spacing:24,className:t.container},n.createElement(r.Hidden,{smDown:!0},n.createElement(r.Grid,{item:!0,xs:3},n.createElement(i.default,null))),n.createElement(r.Grid,{item:!0,xs:12,md:9},n.createElement(r.Paper,{className:t.paper,style:{overflowX:"auto"}},n.createElement(r.Typography,{variant:"h4",paragraph:!0}," Column Model "),n.createElement(r.Divider,null),n.createElement(l.default,null)))))};exports.default=a.withStyles(u)(o);
-},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/","../../components/ColumnModelTable":"AkVR","../../components/DocumentationList":"5NpU"}],"W7b0":[function(require,module,exports) {
-
-},{}],"R7gn":[function(require,module,exports) {
+"use strict";var e=this&&this.__extends||function(){var e=function(t,r){return(e=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var r in t)t.hasOwnProperty(r)&&(e[r]=t[r])})(t,r)};return function(t,r){function n(){this.constructor=t}e(t,r),t.prototype=null===r?Object.create(r):(n.prototype=r.prototype,new n)}}(),t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}},r=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var r in e)Object.hasOwnProperty.call(e,r)&&(t[r]=e[r]);return t.default=e,t};exports.__esModule=!0;var n=require("@material-ui/core"),a=t(require("@material-ui/core/styles/createStyles")),o=t(require("@material-ui/icons/ChevronLeft")),l=t(require("@material-ui/icons/Menu")),i=r(require("react")),s=require("react-router-dom"),c=t(require("./Github")),u=function(){return a.default({flex:{flex:1,textAlign:"center"},image:{maxWidth:100},link:{textDecoration:"none"},logo:{maxWidth:150},nested:{marginLeft:10},spacer:{flex:"1 1 100%"}})},m=n.withStyles(u)(function(t){function r(){var e=null!==t&&t.apply(this,arguments)||this;return e.state={openDrawer:!1,openList:!1},e.toggleDrawer=function(t){return function(){e.setState({openDrawer:t}),e.state.openList&&e.setState({openList:t})}},e.toggleList=function(t){return function(){e.setState({openList:t})}},e}return e(r,t),r.prototype.render=function(){var e=this.props.classes;return i.createElement("div",null,i.createElement(n.AppBar,{position:"static",color:"primary"},i.createElement(n.Toolbar,null,i.createElement(n.Tooltip,{title:"Open"},i.createElement(n.IconButton,{onClick:this.toggleDrawer(!0),color:"inherit"},i.createElement(l.default,null))),i.createElement("img",{className:e.logo,src:"https://unosquare.github.io/tubular-react/static/tubular.png",alt:"Tubular"}),i.createElement("div",{className:e.spacer}),i.createElement(n.Tooltip,{title:"Tubular GitHub Repo"},i.createElement(n.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",className:e.flex,color:"inherit"},i.createElement(c.default,null))))),i.createElement(n.Drawer,{variant:"persistent",open:this.state.openDrawer,onClose:this.toggleDrawer(!1)},i.createElement(n.List,null,i.createElement(n.Tooltip,{title:"Close"},i.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},i.createElement(n.ListItemIcon,null,i.createElement(o.default,null)))),i.createElement(n.Divider,null),i.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},i.createElement(s.Link,{to:"/",className:e.link},i.createElement(n.ListItemText,{primary:"Home"}))),i.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},i.createElement(s.Link,{to:"/Sample",className:e.link},i.createElement(n.ListItemText,{primary:"Grid Samples"}))),i.createElement(n.ListItem,{button:!0,onClick:this.toggleDrawer(!1)},i.createElement(s.Link,{to:"/ComponentAPI",className:e.link},i.createElement(n.ListItemText,{primary:"ComponentAPI"}))),i.createElement(n.Divider,null),i.createElement(n.ListItem,{button:!0,component:"a",href:"https://www.unosquare.com/",target:"_blank",onClick:this.toggleDrawer(!1)},i.createElement("img",{className:e.image,src:"../../static/unosquare_logo.svg",alt:"Unosquare"})))))},r}(i.Component));exports.default=n.withStyles(u)(m);
+},{"@material-ui/core":"J2ra","@material-ui/core/styles/createStyles":"zEWA","@material-ui/icons/ChevronLeft":"KAKw","@material-ui/icons/Menu":"O5EH","react":"1n8/","react-router-dom":"obMO","./Github":"qM49"}],"R7gn":[function(require,module,exports) {
 var define;
 var global = arguments[3];
 var e,n=arguments[3];!function(n){var t="object"==typeof window&&window||"object"==typeof self&&self;"undefined"!=typeof exports?n(exports):t&&(t.hljs=n({}),"function"==typeof e&&e.amd&&e([],function(){return t.hljs}))}(function(e){var n=[],t=Object.keys,a={},r={},i=/^(no-?highlight|plain|text)$/i,s=/\blang(?:uage)?-([\w-]+)\b/i,o=/((^(<[^>]+>|\t|)+|(?:\n)))/gm,l="</span>",c={classPrefix:"hljs-",tabReplace:null,useBR:!1,languages:void 0};function u(e){return e.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;")}function g(e){return e.nodeName.toLowerCase()}function f(e,n){var t=e&&e.exec(n);return t&&0===t.index}function d(e){return i.test(e)}function E(e){var n,t={},a=Array.prototype.slice.call(arguments,1);for(n in e)t[n]=e[n];return a.forEach(function(e){for(n in e)t[n]=e[n]}),t}function v(e){var n=[];return function e(t,a){for(var r=t.firstChild;r;r=r.nextSibling)3===r.nodeType?a+=r.nodeValue.length:1===r.nodeType&&(n.push({event:"start",offset:a,node:r}),a=e(r,a),g(r).match(/br|hr|img|input/)||n.push({event:"stop",offset:a,node:r}));return a}(e,0),n}function m(e){function n(e){return e&&e.source||e}function a(t,a){return new RegExp(n(t),"m"+(e.case_insensitive?"i":"")+(a?"g":""))}!function r(i,s){if(!i.compiled){if(i.compiled=!0,i.keywords=i.keywords||i.beginKeywords,i.keywords){var o={},l=function(n,t){e.case_insensitive&&(t=t.toLowerCase()),t.split(" ").forEach(function(e){var t=e.split("|");o[t[0]]=[n,t[1]?Number(t[1]):1]})};"string"==typeof i.keywords?l("keyword",i.keywords):t(i.keywords).forEach(function(e){l(e,i.keywords[e])}),i.keywords=o}i.lexemesRe=a(i.lexemes||/\w+/,!0),s&&(i.beginKeywords&&(i.begin="\\b("+i.beginKeywords.split(" ").join("|")+")\\b"),i.begin||(i.begin=/\B|\b/),i.beginRe=a(i.begin),i.endSameAsBegin&&(i.end=i.begin),i.end||i.endsWithParent||(i.end=/\B|\b/),i.end&&(i.endRe=a(i.end)),i.terminator_end=n(i.end)||"",i.endsWithParent&&s.terminator_end&&(i.terminator_end+=(i.end?"|":"")+s.terminator_end)),i.illegal&&(i.illegalRe=a(i.illegal)),null==i.relevance&&(i.relevance=1),i.contains||(i.contains=[]),i.contains=Array.prototype.concat.apply([],i.contains.map(function(e){return function(e){return e.variants&&!e.cached_variants&&(e.cached_variants=e.variants.map(function(n){return E(e,{variants:null},n)})),e.cached_variants||e.endsWithParent&&[E(e)]||[e]}("self"===e?i:e)})),i.contains.forEach(function(e){r(e,i)}),i.starts&&r(i.starts,s);var c=i.contains.map(function(e){return e.beginKeywords?"\\.?("+e.begin+")\\.?":e.begin}).concat([i.terminator_end,i.illegal]).map(n).filter(Boolean);i.terminators=c.length?a(c.join("|"),!0):{exec:function(){return null}}}}(e)}function p(e,n,t,r){function i(e,n){var t=E.case_insensitive?n[0].toLowerCase():n[0];return e.keywords.hasOwnProperty(t)&&e.keywords[t]}function s(e,n,t,a){var r='<span class="'+(a?"":c.classPrefix);return(r+=e+'">')+n+(t?"":l)}function o(){_+=null!=b.subLanguage?function(){var e="string"==typeof b.subLanguage;if(e&&!a[b.subLanguage])return u(w);var n=e?p(b.subLanguage,w,!0,R[b.subLanguage]):h(w,b.subLanguage.length?b.subLanguage:void 0);return b.relevance>0&&(M+=n.relevance),e&&(R[b.subLanguage]=n.top),s(n.language,n.value,!1,!0)}():function(){var e,n,t,a;if(!b.keywords)return u(w);for(a="",n=0,b.lexemesRe.lastIndex=0,t=b.lexemesRe.exec(w);t;)a+=u(w.substring(n,t.index)),(e=i(b,t))?(M+=e[1],a+=s(e[0],u(t[0]))):a+=u(t[0]),n=b.lexemesRe.lastIndex,t=b.lexemesRe.exec(w);return a+u(w.substr(n))}(),w=""}function g(e){_+=e.className?s(e.className,"",!0):"",b=Object.create(e,{parent:{value:b}})}function d(e,n){if(w+=e,null==n)return o(),0;var a=function(e,n){var t,a,r;for(t=0,a=n.contains.length;t<a;t++)if(f(n.contains[t].beginRe,e))return n.contains[t].endSameAsBegin&&(n.contains[t].endRe=(r=n.contains[t].beginRe.exec(e)[0],new RegExp(r.replace(/[-\/\\^$*+?.()|[\]{}]/g,"\\$&"),"m"))),n.contains[t]}(n,b);if(a)return a.skip?w+=n:(a.excludeBegin&&(w+=n),o(),a.returnBegin||a.excludeBegin||(w=n)),g(a),a.returnBegin?0:n.length;var r=function e(n,t){if(f(n.endRe,t)){for(;n.endsParent&&n.parent;)n=n.parent;return n}if(n.endsWithParent)return e(n.parent,t)}(b,n);if(r){var i=b;i.skip?w+=n:(i.returnEnd||i.excludeEnd||(w+=n),o(),i.excludeEnd&&(w=n));do{b.className&&(_+=l),b.skip||b.subLanguage||(M+=b.relevance),b=b.parent}while(b!==r.parent);return r.starts&&(r.endSameAsBegin&&(r.starts.endRe=r.endRe),g(r.starts)),i.returnEnd?0:n.length}if(function(e,n){return!t&&f(n.illegalRe,e)}(n,b))throw new Error('Illegal lexeme "'+n+'" for mode "'+(b.className||"<unnamed>")+'"');return w+=n,n.length||1}var E=N(e);if(!E)throw new Error('Unknown language: "'+e+'"');m(E);var v,b=r||E,R={},_="";for(v=b;v!==E;v=v.parent)v.className&&(_=s(v.className,"",!0)+_);var w="",M=0;try{for(var x,O,y=0;b.terminators.lastIndex=y,x=b.terminators.exec(n);)O=d(n.substring(y,x.index),x[0]),y=x.index+O;for(d(n.substr(y)),v=b;v.parent;v=v.parent)v.className&&(_+=l);return{relevance:M,value:_,language:e,top:b}}catch(S){if(S.message&&-1!==S.message.indexOf("Illegal"))return{relevance:0,value:u(n)};throw S}}function h(e,n){n=n||c.languages||t(a);var r={relevance:0,value:u(e)},i=r;return n.filter(N).filter(w).forEach(function(n){var t=p(n,e,!1);t.language=n,t.relevance>i.relevance&&(i=t),t.relevance>r.relevance&&(i=r,r=t)}),i.language&&(r.second_best=i),r}function b(e){return c.tabReplace||c.useBR?e.replace(o,function(e,n){return c.useBR&&"\n"===e?"<br>":c.tabReplace?n.replace(/\t/g,c.tabReplace):""}):e}function R(e){var t,a,i,o,l,f=function(e){var n,t,a,r,i=e.className+" ";if(i+=e.parentNode?e.parentNode.className:"",t=s.exec(i))return N(t[1])?t[1]:"no-highlight";for(n=0,a=(i=i.split(/\s+/)).length;n<a;n++)if(d(r=i[n])||N(r))return r}(e);d(f)||(c.useBR?(t=document.createElementNS("http://www.w3.org/1999/xhtml","div")).innerHTML=e.innerHTML.replace(/\n/g,"").replace(/<br[ \/]*>/g,"\n"):t=e,l=t.textContent,i=f?p(f,l,!0):h(l),(a=v(t)).length&&((o=document.createElementNS("http://www.w3.org/1999/xhtml","div")).innerHTML=i.value,i.value=function(e,t,a){var r=0,i="",s=[];function o(){return e.length&&t.length?e[0].offset!==t[0].offset?e[0].offset<t[0].offset?e:t:"start"===t[0].event?e:t:e.length?e:t}function l(e){i+="<"+g(e)+n.map.call(e.attributes,function(e){return" "+e.nodeName+'="'+u(e.value).replace('"',"&quot;")+'"'}).join("")+">"}function c(e){i+="</"+g(e)+">"}function f(e){("start"===e.event?l:c)(e.node)}for(;e.length||t.length;){var d=o();if(i+=u(a.substring(r,d[0].offset)),r=d[0].offset,d===e){s.reverse().forEach(c);do{f(d.splice(0,1)[0]),d=o()}while(d===e&&d.length&&d[0].offset===r);s.reverse().forEach(l)}else"start"===d[0].event?s.push(d[0].node):s.pop(),f(d.splice(0,1)[0])}return i+u(a.substr(r))}(a,v(o),l)),i.value=b(i.value),e.innerHTML=i.value,e.className=function(e,n,t){var a=n?r[n]:t,i=[e.trim()];return e.match(/\bhljs\b/)||i.push("hljs"),-1===e.indexOf(a)&&i.push(a),i.join(" ").trim()}(e.className,f,i.language),e.result={language:i.language,re:i.relevance},i.second_best&&(e.second_best={language:i.second_best.language,re:i.second_best.relevance}))}function _(){if(!_.called){_.called=!0;var e=document.querySelectorAll("pre code");n.forEach.call(e,R)}}function N(e){return e=(e||"").toLowerCase(),a[e]||a[r[e]]}function w(e){var n=N(e);return n&&!n.disableAutodetect}return e.highlight=p,e.highlightAuto=h,e.fixMarkup=b,e.highlightBlock=R,e.configure=function(e){c=E(c,e)},e.initHighlighting=_,e.initHighlightingOnLoad=function(){addEventListener("DOMContentLoaded",_,!1),addEventListener("load",_,!1)},e.registerLanguage=function(n,t){var i=a[n]=t(e);i.aliases&&i.aliases.forEach(function(e){r[e]=n})},e.listLanguages=function(){return t(a)},e.getLanguage=N,e.autoDetection=w,e.inherit=E,e.IDENT_RE="[a-zA-Z]\\w*",e.UNDERSCORE_IDENT_RE="[a-zA-Z_]\\w*",e.NUMBER_RE="\\b\\d+(\\.\\d+)?",e.C_NUMBER_RE="(-?)(\\b0[xX][a-fA-F0-9]+|(\\b\\d+(\\.\\d*)?|\\.\\d+)([eE][-+]?\\d+)?)",e.BINARY_NUMBER_RE="\\b(0b[01]+)",e.RE_STARTERS_RE="!|!=|!==|%|%=|&|&&|&=|\\*|\\*=|\\+|\\+=|,|-|-=|/=|/|:|;|<<|<<=|<=|<|===|==|=|>>>=|>>=|>=|>>>|>>|>|\\?|\\[|\\{|\\(|\\^|\\^=|\\||\\|=|\\|\\||~",e.BACKSLASH_ESCAPE={begin:"\\\\[\\s\\S]",relevance:0},e.APOS_STRING_MODE={className:"string",begin:"'",end:"'",illegal:"\\n",contains:[e.BACKSLASH_ESCAPE]},e.QUOTE_STRING_MODE={className:"string",begin:'"',end:'"',illegal:"\\n",contains:[e.BACKSLASH_ESCAPE]},e.PHRASAL_WORDS_MODE={begin:/\b(a|an|the|are|I'm|isn't|don't|doesn't|won't|but|just|should|pretty|simply|enough|gonna|going|wtf|so|such|will|you|your|they|like|more)\b/},e.COMMENT=function(n,t,a){var r=e.inherit({className:"comment",begin:n,end:t,contains:[]},a||{});return r.contains.push(e.PHRASAL_WORDS_MODE),r.contains.push({className:"doctag",begin:"(?:TODO|FIXME|NOTE|BUG|XXX):",relevance:0}),r},e.C_LINE_COMMENT_MODE=e.COMMENT("//","$"),e.C_BLOCK_COMMENT_MODE=e.COMMENT("/\\*","\\*/"),e.HASH_COMMENT_MODE=e.COMMENT("#","$"),e.NUMBER_MODE={className:"number",begin:e.NUMBER_RE,relevance:0},e.C_NUMBER_MODE={className:"number",begin:e.C_NUMBER_RE,relevance:0},e.BINARY_NUMBER_MODE={className:"number",begin:e.BINARY_NUMBER_RE,relevance:0},e.CSS_NUMBER_MODE={className:"number",begin:e.NUMBER_RE+"(%|em|ex|ch|rem|vw|vh|vmin|vmax|cm|mm|in|pt|pc|px|deg|grad|rad|turn|s|ms|Hz|kHz|dpi|dpcm|dppx)?",relevance:0},e.REGEXP_MODE={className:"regexp",begin:/\//,end:/\/[gimuy]*/,illegal:/\n/,contains:[e.BACKSLASH_ESCAPE,{begin:/\[/,end:/\]/,relevance:0,contains:[e.BACKSLASH_ESCAPE]}]},e.TITLE_MODE={className:"title",begin:e.IDENT_RE,relevance:0},e.UNDERSCORE_TITLE_MODE={className:"title",begin:e.UNDERSCORE_IDENT_RE,relevance:0},e.METHOD_GUARD={begin:"\\.\\s*"+e.UNDERSCORE_IDENT_RE,relevance:0},e});
@@ -1340,25 +1328,17 @@ var e=require("./highlight");e.registerLanguage("1c",require("./languages/1c")),
 "use strict";Object.defineProperty(exports,"__esModule",{value:!0});var e=function(){function e(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}return function(t,n,r){return n&&e(t.prototype,n),r&&e(t,r),t}}(),t=require("highlight.js"),n=l(t),r=require("react"),o=l(r);function l(e){return e&&e.__esModule?e:{default:e}}function i(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function u(e,t){if(!e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!t||"object"!=typeof t&&"function"!=typeof t?e:t}function a(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function, not "+typeof t);e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),t&&(Object.setPrototypeOf?Object.setPrototypeOf(e,t):e.__proto__=t)}var c=function(t){function r(e){i(this,r);var t=u(this,(r.__proto__||Object.getPrototypeOf(r)).call(this,e));return t.setEl=t.setEl.bind(t),t}return a(r,o.default.Component),e(r,[{key:"componentDidMount",value:function(){this.highlightCode()}},{key:"componentDidUpdate",value:function(){this.highlightCode()}},{key:"highlightCode",value:function(){for(var e=this.el.querySelectorAll("pre code"),t=0;t<e.length;t++)n.default.highlightBlock(e[t])}},{key:"setEl",value:function(e){this.el=e}},{key:"render",value:function(){var e=this.props,t=e.children,n=e.className,r=e.element,l=e.innerHTML,i={ref:this.setEl,className:n};return l?(i.dangerouslySetInnerHTML={__html:t},r?o.default.createElement(r,i):o.default.createElement("div",i)):r?o.default.createElement(r,i,t):o.default.createElement("pre",{ref:this.setEl},o.default.createElement("code",{className:n},t))}}]),r}();c.defaultProps={innerHTML:!1,className:null,element:null},exports.default=c;
 },{"highlight.js":"u5NY","react":"1n8/"}],"Xqwm":[function(require,module,exports) {
 module.exports=require("./lib");
-},{"./lib":"jaG7"}],"EGn3":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var a={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(a[t]=e[t]);return a.default=e,a},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var t=require("@material-ui/core");require("highlight.js/styles/an-old-hope.css");var r=e(require("react")),n=a(require("react-highlight")),o=a(require("../../components/DocumentationList")),c={code:{background:"#F8F8FF",padding:"6px"},container:{margin:"0",padding:30,width:"100%"},paper:{padding:10}},l="... \nconst MyComponent = () => {\n        return (\n            <div className='root'>\n                <DataGrid\n                    gridName='Tubular-React'\n                />\n            </div>\n        );\n};\n\n\n/*\n Use the component withRemoteDataSource to wrap your component\n and columns definition among the data obtained from the URL.\n\n withRemoteDataSource will set an initial context for your grid.\n*/\nexport default withRemoteDataSource(BasicFeatures, columns, 'https://tubular.azurewebsites.net/api/orders/paged');\n",s="... \nconst MyComponent = () => {\n        return (\n            <div className='root'>\n                <DataGrid\n                    gridName='Tubular-React'\n                />\n            </div>\n        );\n};\n\n\nexport default withRemoteDataSource(BasicFeatures, columns, localData)",m="https://github.com/unosquare/tubular-react/blob/master/sample/src/data/localData.ts",i=function(e){var a=e.classes;return r.createElement(t.Grid,{container:!0,spacing:24,className:a.container},r.createElement(t.Hidden,{smDown:!0},r.createElement(t.Grid,{item:!0,xs:3},r.createElement(o.default,null))),r.createElement(t.Grid,{item:!0,xs:12,md:9},r.createElement(t.Paper,{className:a.paper},r.createElement(t.Typography,{variant:"h4",paragraph:!0},"Data Source"),r.createElement(t.Divider,null),r.createElement("br",null),r.createElement(t.Typography,{variant:"subtitle1"},r.createElement("code",{className:a.code},"<DataGrid />")," requires a wrapper component to retrieve data. Depends on if we use a local data source or a remote data source, these wrappers are:",r.createElement("code",{className:a.code},"withRemoteDataSource"),"component or the",r.createElement("code",{className:a.code},"withLocalDataSource"),"component, which deals with data retrieval on the",r.createElement("code",{className:a.code},"DataGrid")," component.",r.createElement("br",null),r.createElement("br",null),r.createElement(t.Typography,{variant:"h5",gutterBottom:!0},r.createElement("code",{className:a.code},"withRemoteDataSource")),r.createElement("code",{className:a.code},"withRemoteDataSource"),"needs our ",r.createElement("code",{className:a.code},"DataGrid"),"component, a ",r.createElement("code",{className:a.code},"ColumnModel"),"array and a ",r.createElement("code",{className:a.code},"URL")," which represent a service.",r.createElement("br",null),r.createElement(n.default,{language:"javascript",className:"an-old-hope"},l),r.createElement("br",null),r.createElement(t.Typography,{variant:"h5",gutterBottom:!0},r.createElement("code",{className:a.code},"LocalDataSource")),r.createElement("code",{className:a.code},"withLocalDataSource"),"needs our ",r.createElement("code",{className:a.code},"DataGrid"),"component, a ",r.createElement("code",{className:a.code},"ColumnModel"),"array and an ",r.createElement("code",{className:a.code},"Array"),"of objects.",r.createElement("br",null),r.createElement(n.default,{language:"javascript",className:"an-old-hope"},s),"See this",r.createElement(t.Button,{color:"primary",href:m,target:"_blank"},"example"),"of how to define the array of objects."))))};exports.default=t.withStyles(c)(i);
-},{"@material-ui/core":"J2ra","highlight.js/styles/an-old-hope.css":"W7b0","react":"1n8/","react-highlight":"Xqwm","../../components/DocumentationList":"5NpU"}],"Gxw2":[function(require,module,exports) {
-"use strict";exports.__esModule=!0,exports.quickStart="import React from 'react';\nimport ReactDOM from 'react-dom';\n\nimport DataGrid, { withRemoteDataSource } from 'tubular-react';\nimport { ColumnModel } from 'tubular-common';\nimport { IconButton, Tooltip } from '@material-ui/core';\n\nconst columns = [\n    new ColumnModel('OrderID'),\n    new ColumnModel('CustomerName'),\n    new ColumnModel('ShipperCity')\n];\n\nconst SampleGrid = withRemoteDataSource(\n    () => {\n        return <DataGrid />;\n    },\n    columns,\n    'https://tubular.azurewebsites.net/api/orders/paged'\n);\n\nReactDOM.render(<SampleGrid />, document.getElementById('root'));",exports.basicFeatures="import * as React from 'react';\n\nimport DataGrid, {\n    withRemoteDataSource } from 'tubular-react';\n\nimport { AggregateFunctions, ColumnDataType, ColumnModel, ColumnSortDirection } from 'tubular-common';\n\n//  First of all, you must define your columns model.\nconst columns = [\n    new ColumnModel('OrderID',\n        {\n            DataType: ColumnDataType.NUMERIC,\n            Filterable: true,\n            IsKey: true,\n            Label: 'ID',\n            SortDirection: ColumnSortDirection.ASCENDING,\n            SortOrder: 1,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('CustomerName',\n        {\n            Aggregate: AggregateFunctions.COUNT,\n            Filterable: true,\n            Searchable: true,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('ShippedDate',\n        {\n            DataType: ColumnDataType.DATE_TIME,\n            Filterable: true,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('ShipperCity'),\n    new ColumnModel('Amount',\n        {\n            DataType: ColumnDataType.NUMERIC,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('IsShipped',\n        {\n            DataType: ColumnDataType.BOOLEAN,\n            Filterable: true,\n            Sortable: true\n        }\n    )\n];\nconst MyComponent = () => {\n\n      <DataGrid gridName='Tubular-React'/>\n};\n\n/*\n Use the component withRemoteDataSource to wrap your component\n and columns definition among the data obtained from the URL.\n\n withRemoteDataSource will set an initial context for your grid.\n*/\nexport default withRemoteDataSource(MyComponent, columns, 'http://tubular.azurewebsites.net/api/orders/paged');",exports.simpleFeatures="import React from 'react';\nimport { Snackbar, TableCell, TableRow } from '@material-ui/core';\nimport CheckBox from '@material-ui/icons/CheckBox';\nimport CheckBoxOutlineBlank from '@material-ui/icons/CheckBoxOutlineBlank';\nimport { format } from 'date-fns';\nimport DataGrid, { ToolbarOptions, withRemoteDataSource } from 'tubular-react';\n\nimport {\n  AggregateFunctions,\n  ColumnDataType,\n  ColumnModel,\n  ColumnSortDirection\n} from 'tubular-common';\n\nconst columns = [\n  new ColumnModel('OrderID', {\n    DataType: ColumnDataType.NUMERIC,\n    Filterable: true,\n    IsKey: true,\n    Label: 'ID',\n    SortDirection: ColumnSortDirection.ASCENDING,\n    SortOrder: 1,\n    Sortable: true\n  }),\n  new ColumnModel('CustomerName', {\n    Aggregate: AggregateFunctions.COUNT,\n    Filterable: true,\n    Searchable: true,\n    Sortable: true\n  }),\n  new ColumnModel('ShippedDate', {\n    DataType: ColumnDataType.DATE_TIME,\n    Filterable: true,\n    Sortable: true\n  }),\n  new ColumnModel('ShipperCity'),\n  new ColumnModel('Amount', {\n    DataType: ColumnDataType.NUMERIC,\n    Sortable: true\n  }),\n  new ColumnModel('IsShipped', {\n    DataType: ColumnDataType.BOOLEAN,\n    Filterable: true,\n    Sortable: true\n  })\n];\n\nclass SampleFeatures extends React.Component {\n  state = {\n    errorMessage: null\n  };\n\n  componentWillReceiveProps(nextProps: any) {\n    this.setState({ errorMessage: nextProps.error });\n  }\n\n  render() {\n    const { errorMessage } = this.state;\n\n    return (\n      <div className='root'>\n        {errorMessage && (\n          <Snackbar\n            anchorOrigin={{ vertical: 'top', horizontal: 'center' }}\n            style={{ paddingTop: '10px' }}\n            open={true}\n            ContentProps={{ 'aria-describedby': 'message-id' }}\n            message={<span id='message-id'>{errorMessage}</span>}\n          />\n        )}\n        <DataGrid\n          gridName='Tubular-React'\n          bodyRenderer={(row: any, index: number, columns: ColumnModel[]) => (\n            <TableRow hover={true} key={row}>\n              <TableCell padding='default'>{row.OrderID}</TableCell>\n              <TableCell padding='default'>{row.CustomerName}</TableCell>\n              <TableCell padding='default'>\n                {format(row.ShippedDate, 'MMMM Do YYYY, h:mm:ss a')}\n              </TableCell>\n              <TableCell padding='default'>{row.ShipperCity}</TableCell>\n              <TableCell padding='default' numeric={true}>\n                {row.Amount || 0}\n              </TableCell>\n              <TableCell padding='default'>\n                {row.IsShipped ? <CheckBox /> : <CheckBoxOutlineBlank />}\n              </TableCell>\n            </TableRow>\n          )}\n          toolbarOptions={new ToolbarOptions()}\n          footerRenderer={(aggregates: any) => (\n            <TableRow>\n              <TableCell>Total: </TableCell>\n              <TableCell>{aggregates && aggregates.CustomerName}</TableCell>\n              <TableCell />\n              <TableCell />\n              <TableCell />\n            </TableRow>\n          )}\n        />\n      </div>\n    );\n  }\n}\n\nexport default withRemoteDataSource(\n  SampleFeatures,\n  columns,\n  'https://tubular.azurewebsites.net/api/orders/paged'\n);",exports.gridList="import React from 'react';\nimport ReactDOM from 'react-dom';\nimport { LinearProgress, Paper } from '@material-ui/core';\nimport { Card, CardActions, CardContent } from '@material-ui/core';\nimport { GridList, GridListTile, Typography } from '@material-ui/core';\nimport { Button } from '@material-ui/core';\nimport {\n  Table,\n  TableBody,\n  TableCell,\n  TableFooter,\n  TableRow\n} from '@material-ui/core';\nimport { format } from 'date-fns';\nimport {\n  DataSourceContext,\n  Paginator,\n  TextSearchInput,\n  withRemoteDataSource,\n  IDataGridState,\n  IDataGridProps\n} from 'tubular-react';\nimport {\n  AggregateFunctions,\n  ColumnDataType,\n  ColumnModel,\n  ColumnSortDirection\n} from 'tubular-common';\n\nconst columns = [\n  new ColumnModel('OrderID', {\n    DataType: ColumnDataType.NUMERIC,\n    Filterable: true,\n    IsKey: true,\n    Label: 'ID',\n    SortDirection: ColumnSortDirection.ASCENDING,\n    SortOrder: 1,\n    Sortable: true\n  }),\n  new ColumnModel('CustomerName', {\n    Aggregate: AggregateFunctions.COUNT,\n    Filterable: true,\n    Searchable: true,\n    Sortable: true\n  }),\n  new ColumnModel('ShippedDate', {\n    DataType: ColumnDataType.DATE_TIME,\n    Filterable: true,\n    Sortable: true\n  }),\n  new ColumnModel('ShipperCity'),\n  new ColumnModel('Amount', {\n    DataType: ColumnDataType.NUMERIC,\n    Sortable: true\n  }),\n  new ColumnModel('IsShipped', {\n    DataType: ColumnDataType.BOOLEAN,\n    Filterable: true,\n    Sortable: true\n  })\n];\n\nconst styles: any = {\n  progress: {\n    height: '20px'\n  },\n  search: {\n    margin: '15px 10px 10px 10px',\n    textAlign: 'right'\n  }\n};\n\nclass SampleGridList extends React.Component<IDataGridProps, IDataGridState> {\n  state = { errorMessage: null };\n\n\n  static getDerivedStateFromProps(\n    props: IDataGridProps,\n    state: IDataGridState\n  ) {\n    if (props.error !== state.errorMessage) {\n      return { errorMessage: props.error };\n    }\n    return null;\n  }\n\n  render() {\n    return (\n      <DataSourceContext.Consumer>\n        {({ state }) => (\n          <Paper>\n            <div style={styles.search}>\n              <TextSearchInput />\n            </div>\n            <div style={styles.progress}>\n              {state.isLoading && <LinearProgress />}\n            </div>\n            <Table>\n              <TableBody>\n                <TableRow>\n                  <TableCell>\n                    <GridList cellHeight={180} cols={5}>\n                      {state.data.map(item => (\n                        <GridListTile key={item.OrderID}>\n                          <Card>\n                            <CardContent>\n                              <Typography\n                                gutterBottom={true}\n                                variant='headline'\n                                component='h2'\n                              >\n                                {item.OrderID} - {item.CustomerName}\n                              </Typography>\n                              <Typography component='p'>\n                                {item.ShipperCity}\n                              </Typography>\n                              <Typography component='p'>\n                                {format(item.ShippedDate, 'MMM D YYYY')}\n                              </Typography>\n                            </CardContent>\n                            <CardActions>\n                              <Button size='small' color='primary'>\n                                Learn More\n                              </Button>\n                            </CardActions>\n                          </Card>\n                        </GridListTile>\n                      ))}\n                    </GridList>\n                  </TableCell>\n                </TableRow>\n              </TableBody>\n              <TableFooter>\n                <TableRow>\n                  <Paginator advancePagination={false} />\n                </TableRow>\n              </TableFooter>\n            </Table>\n          </Paper>\n        )}\n      </DataSourceContext.Consumer>\n    );\n  }\n}\n\nexport default withRemoteDataSource(\n  SampleGridList,\n  columns,\n  'https://tubular.azurewebsites.net/api/orders/paged'\n);";
-},{}],"VWQ5":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var a={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(a[t]=e[t]);return a.default=e,a},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var t=require("@material-ui/core"),r=require("@material-ui/core/styles");require("highlight.js/styles/an-old-hope.css");var n=e(require("react")),l=a(require("react-highlight")),i=a(require("../../components/DocumentationList")),c=require("../../utils/codeSnipetExamples"),o={code:{fontSize:15},codeTag:{background:"#F8F8FF",padding:"6px"},container:{margin:"0",padding:30,width:"100%"},paper:{padding:10}},s=function(e){var a=e.classes;return n.createElement(t.Grid,{container:!0,spacing:24,className:a.container},n.createElement(t.Hidden,{smDown:!0},n.createElement(t.Grid,{item:!0,xs:3},n.createElement(i.default,null))),n.createElement(t.Grid,{item:!0,xs:12,md:9},n.createElement(t.Paper,{className:a.paper},n.createElement(t.Typography,{variant:"h4",paragraph:!0},"Getting Started"),n.createElement(t.Divider,null),n.createElement("br",null),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"Installation"),n.createElement(t.Typography,{variant:"subtitle1",paragraph:!0},"Tubular-React is available as an",n.createElement("a",{href:"https://www.npmjs.com/package/tubular-react",target:"_blank"},"npm package")),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"Dependencies"),n.createElement(t.Typography,{variant:"body1",paragraph:!0},n.createElement("ul",null,n.createElement("li",null,n.createElement("a",{href:"https://date-fns.org/"},"date-fns - Version: 1.29.0")),n.createElement("li",null,n.createElement("a",{href:"https://material-ui.com/"},"Material-UI - Version: 3.1.0")),n.createElement("li",null,n.createElement("a",{href:"https://reactjs.org/"},"React - Version: 16.5.1")),n.createElement("li",null,n.createElement("a",{href:"https://github.com/unosquare/tubular-common"},"Tubular Common - Version: 1.2.1")))),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"npm"),n.createElement(t.Typography,{variant:"body1",paragraph:!0},"To install and save in your",n.createElement("code",{className:a.codeTag},"package.json"),"dependencies, run:"),n.createElement(l.default,{language:"javascript",className:a.code},"npm install tubular-react --save"),n.createElement("br",null),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"Usage"),n.createElement(t.Typography,{variant:"subtitle1",paragraph:!0},"Tubular React is an extension of Material-UI which offers until now a couple of useful components:",n.createElement("br",null),n.createElement("code",{className:a.codeTag},"<DataGrid />")," and",n.createElement("code",{className:a.codeTag},"<GridList />"),".",n.createElement("br",null),"These components supply an organized and nice way to represent your data, you can populate them from a server-side data source or a local data source as an array."),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"Quick Start"),n.createElement(t.Typography,{variant:"body1",paragraph:!0},"Here is a quick example with all basic features to get you started:"),n.createElement(l.default,{language:"javascript",className:a.code},c.basicFeatures),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"Extend Grid Functionalities"),n.createElement(t.Typography,{variant:"subtitle1",paragraph:!0},"You can add functionalities to the `DataGrid`, including extra buttons that can perform an action according to your requirements. Just need include an IconButton Component from @material-ui and define the icon or button that you need between `DataGrid` tags and specify the action to perform.",n.createElement("br",null),n.createElement(l.default,{language:"javascript",className:a.code},"\n              const MyComponent = () => {\n\n              <DataGrid gridName='Tubular-React'>\n              <IconButton color='default' />\n              <Brightness7Rounded\n                onClick={() => alert('Happy codes, have a nice day')}\n                />\n              </IconButton>\n              </DataGrid>\n\n            };\n              ")))))};exports.default=r.withStyles(o)(s);
-},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","highlight.js/styles/an-old-hope.css":"W7b0","react":"1n8/","react-highlight":"Xqwm","../../components/DocumentationList":"5NpU","../../utils/codeSnipetExamples":"Gxw2"}],"9PWw":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var a in e)Object.hasOwnProperty.call(e,a)&&(t[a]=e[a]);return t.default=e,t},t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var a=require("@material-ui/core"),l=require("@material-ui/core/styles"),r=e(require("react")),n=t(require("react-highlight")),i=require("../../utils/codeSnipetExamples"),o={code:{fontSize:15},codeTag:{background:"#F8F8FF",padding:"6px"},container:{margin:"0",padding:30,width:"100%"},content:{marginTop:10},image:{marginLeft:"25%",maxWidth:300},paper:{minHeight:"300px",padding:30},paperSmall:{minHeight:"115px",padding:30}},s=function(e){var t=e.classes;return r.createElement(a.Grid,{container:!0,spacing:24,className:t.container},r.createElement(a.Paper,{className:t.paper},r.createElement(a.Typography,{variant:"h4"},"What is it?"),r.createElement(a.Divider,null),r.createElement(a.Typography,{paragraph:!0,variant:"subtitle1",className:t.content},"Tubular React is a set of ReactJS components designed to rapidly build modern web applications. The centerpiece of Tubular is its fully templated grid with lots of features such as server-side pagination, searching text, multi-column sorting, and Filterable, built-in export to CSV or been printed. Another very styled and useful component is the grid list that renders a set of cards with the general information allowing good data visualization and quickly searching, besides it has almost the same grid functionalities."),r.createElement("br",null),r.createElement(a.Typography,{variant:"h4"},"Features"),r.createElement(a.Divider,null),r.createElement(a.Typography,{paragraph:!0,variant:"subtitle1",className:t.content},r.createElement(a.Typography,{variant:"h5"},"The main component is a ",r.createElement("em",null,"grid")," with multiple options:"),r.createElement("ul",null,r.createElement("li",null,"Define a custom layout for columns and cells using render methods."),r.createElement("li",null,"Use a remote or local datasource. Remote datasource use a specific Request and Response format."),r.createElement("li",null,"Sort and filter multiple columns."),r.createElement("li",null,"Free-text search of string columns."),r.createElement("li",null,"Page data. Remote data is paged in the server side."),r.createElement("li",null,"Export data to a CSV file."),r.createElement("li",null,"Print data.")),r.createElement("br",null),"The ",r.createElement("em",null,"grid")," component offers all you can look for an easy grid table and more.",r.createElement("br",null),r.createElement("br",null),r.createElement(a.Typography,{variant:"h5"},r.createElement("em",null,"Grid list")," component allows you to perform some of the same options as:"),r.createElement("ul",null,r.createElement("li",null,"Define a custom layout for columns and cells."),r.createElement("li",null,"Use a remote or local datasource."),r.createElement("li",null,"Free-text search of string columns."),r.createElement("li",null,"Page data. Remote data is paged in the server side.")),r.createElement("br",null),"The ",r.createElement("em",null,"grid list")," component represents another option to show your data.",r.createElement("br",null),r.createElement("em",null,"Grid list")," provide a styled and fancy view over your data, which helps you to quickly find information."),r.createElement("br",null),r.createElement(a.Typography,{variant:"h4"},"Dependencies"),r.createElement(a.Divider,null),r.createElement(a.Typography,{paragraph:!0,variant:"subtitle1"},r.createElement("ul",null,r.createElement("li",null,r.createElement("a",{href:"https://date-fns.org/"},"date-fns - Version: 1.29.0")),r.createElement("li",null,r.createElement("a",{href:"https://material-ui.com/"},"Material-UI - Version: 3.1.0")),r.createElement("li",null,r.createElement("a",{href:"https://reactjs.org/"},"React - Version: 16.5.1")),r.createElement("li",null,r.createElement("a",{href:"https://github.com/unosquare/tubular-common"},"Tubular Common - Version: 1.2.1")))),r.createElement("br",null),r.createElement(a.Typography,{variant:"h4"},"npm Installation"),r.createElement(a.Divider,null),r.createElement(a.Typography,{paragraph:!0,variant:"subtitle1",className:t.content},r.createElement(n.default,{language:"javascript",className:t.code},"$ npm install tubular-react --save")),r.createElement("br",null),r.createElement(a.Typography,{variant:"h4"},"Usage"),r.createElement(a.Divider,null),r.createElement(a.Typography,{paragraph:!0,variant:"subtitle1",className:t.content},"Here is a quick example to get you started, ",r.createElement("b",null,"it's all you need"),":",r.createElement(n.default,{language:"javascript",className:t.code},i.quickStart),r.createElement("em",null,"Try it in",r.createElement("a",{href:"https://codesandbox.io/s/3rqq6l04jq",target:"_blank"},"CodeSandbox")))))};exports.default=l.withStyles(o)(s);
-},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/","react-highlight":"Xqwm","../../utils/codeSnipetExamples":"Gxw2"}],"WKKG":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var l={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(l[t]=e[t]);return l.default=e,l};exports.__esModule=!0;var l=require("@material-ui/core"),t=require("@material-ui/core/styles"),a=e(require("react")),n=function(e){var l=e.spacing;return t.createStyles({code:{background:"#F8F8FF",fontSize:14},root:{marginTop:3*l.unit,overflowX:"auto",width:"100%"}})},r=function(e){var t=e.classes;return a.createElement(l.Paper,{className:t.root},a.createElement(l.Table,null,a.createElement(l.TableHead,null,a.createElement(l.TableRow,null,a.createElement(l.TableCell,null,"Name"),a.createElement(l.TableCell,null,"Type"),a.createElement(l.TableCell,null,"Default value"),a.createElement(l.TableCell,null,"Description"))),a.createElement(l.TableBody,null,a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"gridName")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"string")),a.createElement(l.TableCell,null,"Grid"),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"toolbarOptions")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"ToolbarOptions")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"new ToolBarOptions();")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional."),"It should be an instance of ToolbarOptions. This encapsulates useful options.",a.createElement("em",null,"see ToolBarOptions props"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bodyRenderer")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"function")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional."),"It takes a function with 3 parameters",a.createElement("code",{className:t.code},"(row: any, index: number, columns: ColumnModel[])")," to map all rows.",a.createElement("em",null,"see sample Full grid features"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"footerRenderer")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"function")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional."),"It takes an aggregate function with 1 parameter",a.createElement("code",{className:t.code},"(aggregate: any)")," to show in a foot row the results from the aggregate function.",a.createElement("em",null,"see sample Full grid features"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"onRowClick")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"function")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional."),"Use this event handler to receive the clicked row.")))))};exports.default=t.withStyles(n)(r);
-},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/"}],"AtZZ":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var l={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(l[t]=e[t]);return l.default=e,l};exports.__esModule=!0;var l=require("@material-ui/core"),t=require("@material-ui/core/styles"),a=e(require("react")),n=function(e){return t.createStyles({code:{background:"#F8F8FF",fontSize:14},root:{marginTop:3*e.spacing.unit,overflowX:"auto",width:"100%"}})},r=function(e){var t=e.classes;return a.createElement(l.Paper,{className:t.root},a.createElement(l.Table,null,a.createElement(l.TableHead,null,a.createElement(l.TableRow,null,a.createElement(l.TableCell,null,"Name"),a.createElement(l.TableCell,null,"Type"),a.createElement(l.TableCell,null,"Default value"),a.createElement(l.TableCell,null,"Description"))),a.createElement(l.TableBody,null,a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"advancePagination")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows advanced pagination with numbers of the page to navigate between pages, also display the navigation buttons to the last or first page. Otherwise, show a simple pagination with two directions arrows for advance or go back a page.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bottomPager")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows pagination bar in the bottom of the grid.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"exportButton")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows the export button.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"printButton")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows the print button.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"searchText")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows the search text input.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"topPager")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows pagination bar in the top of the grid.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"rowsPerPageOptions")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"array")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"[10, 20, 50, 100]")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows a combo with a set of values that represent the number of rows per page.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"rowsPerPage")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"number")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"10")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"Initial valor that represent the number of rows per page. It should be a number that is inside the ",a.createElement("em",null,"rowsPerPageOptions")," array.")))))};exports.default=t.withStyles(n)(r);
-},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/"}],"V9e7":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var a in e)Object.hasOwnProperty.call(e,a)&&(t[a]=e[a]);return t.default=e,t},t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var a=require("@material-ui/core"),r=require("@material-ui/core/styles"),l=e(require("react")),n=t(require("../../components/DataGridProps")),o=t(require("../../components/DocumentationList")),i=t(require("../../components/ToolBarOptionsProps")),c={code:{background:"#F8F8FF",padding:"6px"},container:{margin:"0",padding:30,width:"100%"},paper:{padding:10}},s=function(e){var t=e.classes;return l.createElement(a.Grid,{container:!0,spacing:24,className:t.container},l.createElement(a.Hidden,{smDown:!0},l.createElement(a.Grid,{item:!0,xs:3},l.createElement(o.default,null))),l.createElement(a.Grid,{item:!0,xs:12,md:9},l.createElement(a.Paper,{className:t.paper},l.createElement(a.Typography,{variant:"h4",paragraph:!0},"DataGrid props"),l.createElement(a.Divider,null),l.createElement("br",null),l.createElement(a.Typography,{variant:"subtitle1"},"It's important to use ",l.createElement("code",{className:t.code},"withRemoteDataSource")," or",l.createElement("code",{className:t.code},"withLocalDataSource")," according to the case to fill the ",l.createElement("code",{className:t.code},"<DataGrid />")," component with the data.",l.createElement("br",null),"These are all the available props (and their default values) for the",l.createElement("code",{className:t.code},"<DataGrid />")," component."),l.createElement(n.default,null),l.createElement("br",null),l.createElement(a.Typography,{variant:"subtitle1"},l.createElement("i",null,"If you don't define some of the optional props described above, these will not be shown. In the case of "),l.createElement("code",{className:t.code},"bodyRenderer"),",",l.createElement("i",null,"the grid will display its default body; if the",l.createElement("code",{className:t.code},"footerRenderer"),"is not defined, the footer will not be displayed."))),l.createElement("br",null),l.createElement(a.Paper,{className:t.paper},l.createElement(a.Typography,{variant:"h4",paragraph:!0},"Toolbar options"),l.createElement(a.Divider,null),l.createElement("br",null),l.createElement(a.Typography,{variant:"subtitle1"},"If you need personalite the grid adding, removing or modifying  fetures, ",l.createElement("code",{className:t.code},"ToolBarOptions Class"),"provides several options."),l.createElement(i.default,null),l.createElement("br",null),l.createElement(a.Typography,{variant:"subtitle1"},l.createElement("i",null,"If you don't define some of the optional props described above, these will set with the default values.")))))};exports.default=r.withStyles(c)(s);
-},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/","../../components/DataGridProps":"WKKG","../../components/DocumentationList":"5NpU","../../components/ToolBarOptionsProps":"AtZZ"}],"vRtn":[function(require,module,exports) {
+},{"./lib":"jaG7"}],"Gxw2":[function(require,module,exports) {
+"use strict";exports.__esModule=!0,exports.quickStart="import React from 'react';\nimport ReactDOM from 'react-dom';\n\nimport DataGrid, { withRemoteDataSource } from 'tubular-react';\nimport { ColumnModel } from 'tubular-common';\nimport { IconButton, Tooltip } from '@material-ui/core';\n\nconst columns = [\n    new ColumnModel('OrderID'),\n    new ColumnModel('CustomerName'),\n    new ColumnModel('ShipperCity')\n];\n\nconst SampleGrid = withRemoteDataSource(\n    () => {\n        return <DataGrid />;\n    },\n    columns,\n    'https://tubular.azurewebsites.net/api/orders/paged'\n);\n\nReactDOM.render(<SampleGrid />, document.getElementById('root'));",exports.basicFeatures="import * as React from 'react';\n\nimport DataGrid, {\n    withRemoteDataSource } from 'tubular-react';\n\nimport { AggregateFunctions, ColumnDataType, ColumnModel, ColumnSortDirection } from 'tubular-common';\n\n//  First of all, you must define your columns model.\nconst columns = [\n    new ColumnModel('OrderID',\n        {\n            DataType: ColumnDataType.NUMERIC,\n            Filterable: true,\n            IsKey: true,\n            Label: 'ID',\n            SortDirection: ColumnSortDirection.ASCENDING,\n            SortOrder: 1,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('CustomerName',\n        {\n            Aggregate: AggregateFunctions.COUNT,\n            Filterable: true,\n            Searchable: true,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('ShippedDate',\n        {\n            DataType: ColumnDataType.DATE_TIME,\n            Filterable: true,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('ShipperCity'),\n    new ColumnModel('Amount',\n        {\n            DataType: ColumnDataType.NUMERIC,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('IsShipped',\n        {\n            DataType: ColumnDataType.BOOLEAN,\n            Filterable: true,\n            Sortable: true\n        }\n    )\n];\nconst MyComponent = () => {\n\n      <DataGrid gridName='Tubular-React'/>\n};\n\n/*\n Use the component withRemoteDataSource to wrap your component\n and columns definition among the data obtained from the URL.\n\n withRemoteDataSource will set an initial context for your grid.\n*/\nexport default withRemoteDataSource(MyComponent, columns, 'http://tubular.azurewebsites.net/api/orders/paged');",exports.buttonFeatures="import * as React from 'react';\nimport IconButton from '@material-ui/core/IconButton';\nimport Brightness7Rounded from '@material-ui/icons/Brightness7Rounded';\nimport Mood from '@material-ui/icons/Mood';\n\nimport DataGrid, {\n    withRemoteDataSource } from 'tubular-react';\n\nimport { AggregateFunctions, ColumnDataType, ColumnModel, ColumnSortDirection } from 'tubular-common';\n\n//  First of all, you must define your columns model.\nconst columns = [\n    new ColumnModel('OrderID',\n        {\n            DataType: ColumnDataType.NUMERIC,\n            Filterable: true,\n            IsKey: true,\n            Label: 'ID',\n            SortDirection: ColumnSortDirection.ASCENDING,\n            SortOrder: 1,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('CustomerName',\n        {\n            Aggregate: AggregateFunctions.COUNT,\n            Filterable: true,\n            Searchable: true,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('ShippedDate',\n        {\n            DataType: ColumnDataType.DATE_TIME,\n            Filterable: true,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('ShipperCity'),\n    new ColumnModel('Amount',\n        {\n            DataType: ColumnDataType.NUMERIC,\n            Sortable: true\n        }\n    ),\n    new ColumnModel('IsShipped',\n        {\n            DataType: ColumnDataType.BOOLEAN,\n            Filterable: true,\n            Sortable: true\n        }\n    )\n];\nconst MyComponent = () => {\n\n      <DataGrid gridName='Tubular-React'>\n        <IconButton color='default'>\n          <Mood\n            onClick={() =>\n              alert('I can help you to add features to your datagrid.')\n            }\n          />\n        </IconButton>\n        <IconButton color='default'>\n          <Brightness7Rounded\n            onClick={() => alert('Happy codes, have a nice day')}\n          />\n        </IconButton>\n      </DataGrid>\n};\n\n/*\n Use the component withRemoteDataSource to wrap your component\n and columns definition among the data obtained from the URL.\n\n withRemoteDataSource will set an initial context for your grid.\n*/\nexport default withRemoteDataSource(MyComponent, columns, 'http://tubular.azurewebsites.net/api/orders/paged');",exports.simpleFeatures="import React from 'react';\nimport { Snackbar, TableCell, TableRow } from '@material-ui/core';\nimport CheckBox from '@material-ui/icons/CheckBox';\nimport CheckBoxOutlineBlank from '@material-ui/icons/CheckBoxOutlineBlank';\nimport { format } from 'date-fns';\nimport DataGrid, { ToolbarOptions, withRemoteDataSource } from 'tubular-react';\n\nimport {\n  AggregateFunctions,\n  ColumnDataType,\n  ColumnModel,\n  ColumnSortDirection\n} from 'tubular-common';\n\nconst columns = [\n  new ColumnModel('OrderID', {\n    DataType: ColumnDataType.NUMERIC,\n    Filterable: true,\n    IsKey: true,\n    Label: 'ID',\n    SortDirection: ColumnSortDirection.ASCENDING,\n    SortOrder: 1,\n    Sortable: true\n  }),\n  new ColumnModel('CustomerName', {\n    Aggregate: AggregateFunctions.COUNT,\n    Filterable: true,\n    Searchable: true,\n    Sortable: true\n  }),\n  new ColumnModel('ShippedDate', {\n    DataType: ColumnDataType.DATE_TIME,\n    Filterable: true,\n    Sortable: true\n  }),\n  new ColumnModel('ShipperCity'),\n  new ColumnModel('Amount', {\n    DataType: ColumnDataType.NUMERIC,\n    Sortable: true\n  }),\n  new ColumnModel('IsShipped', {\n    DataType: ColumnDataType.BOOLEAN,\n    Filterable: true,\n    Sortable: true\n  })\n];\n\nclass SampleFeatures extends React.Component {\n  state = {\n    errorMessage: null\n  };\n\n  componentWillReceiveProps(nextProps: any) {\n    this.setState({ errorMessage: nextProps.error });\n  }\n\n  render() {\n    const { errorMessage } = this.state;\n\n    return (\n      <div className='root'>\n        {errorMessage && (\n          <Snackbar\n            anchorOrigin={{ vertical: 'top', horizontal: 'center' }}\n            style={{ paddingTop: '10px' }}\n            open={true}\n            ContentProps={{ 'aria-describedby': 'message-id' }}\n            message={<span id='message-id'>{errorMessage}</span>}\n          />\n        )}\n        <DataGrid\n          gridName='Tubular-React'\n          bodyRenderer={(row: any, index: number, columns: ColumnModel[]) => (\n            <TableRow hover={true} key={row}>\n              <TableCell padding='default'>{row.OrderID}</TableCell>\n              <TableCell padding='default'>{row.CustomerName}</TableCell>\n              <TableCell padding='default'>\n                {format(row.ShippedDate, 'MMMM Do YYYY, h:mm:ss a')}\n              </TableCell>\n              <TableCell padding='default'>{row.ShipperCity}</TableCell>\n              <TableCell padding='default' numeric={true}>\n                {row.Amount || 0}\n              </TableCell>\n              <TableCell padding='default'>\n                {row.IsShipped ? <CheckBox /> : <CheckBoxOutlineBlank />}\n              </TableCell>\n            </TableRow>\n          )}\n          toolbarOptions={new ToolbarOptions()}\n          footerRenderer={(aggregates: any) => (\n            <TableRow>\n              <TableCell>Total: </TableCell>\n              <TableCell>{aggregates && aggregates.CustomerName}</TableCell>\n              <TableCell />\n              <TableCell />\n              <TableCell />\n            </TableRow>\n          )}\n        />\n      </div>\n    );\n  }\n}\n\nexport default withRemoteDataSource(\n  SampleFeatures,\n  columns,\n  'https://tubular.azurewebsites.net/api/orders/paged'\n);",exports.gridList="import React from 'react';\nimport ReactDOM from 'react-dom';\nimport { LinearProgress, Paper } from '@material-ui/core';\nimport { Card, CardActions, CardContent } from '@material-ui/core';\nimport { GridList, GridListTile, Typography } from '@material-ui/core';\nimport { Button } from '@material-ui/core';\nimport {\n  Table,\n  TableBody,\n  TableCell,\n  TableFooter,\n  TableRow\n} from '@material-ui/core';\nimport { format } from 'date-fns';\nimport {\n  DataSourceContext,\n  Paginator,\n  TextSearchInput,\n  withRemoteDataSource,\n  IDataGridState,\n  IDataGridProps\n} from 'tubular-react';\nimport {\n  AggregateFunctions,\n  ColumnDataType,\n  ColumnModel,\n  ColumnSortDirection\n} from 'tubular-common';\n\nconst columns = [\n  new ColumnModel('OrderID', {\n    DataType: ColumnDataType.NUMERIC,\n    Filterable: true,\n    IsKey: true,\n    Label: 'ID',\n    SortDirection: ColumnSortDirection.ASCENDING,\n    SortOrder: 1,\n    Sortable: true\n  }),\n  new ColumnModel('CustomerName', {\n    Aggregate: AggregateFunctions.COUNT,\n    Filterable: true,\n    Searchable: true,\n    Sortable: true\n  }),\n  new ColumnModel('ShippedDate', {\n    DataType: ColumnDataType.DATE_TIME,\n    Filterable: true,\n    Sortable: true\n  }),\n  new ColumnModel('ShipperCity'),\n  new ColumnModel('Amount', {\n    DataType: ColumnDataType.NUMERIC,\n    Sortable: true\n  }),\n  new ColumnModel('IsShipped', {\n    DataType: ColumnDataType.BOOLEAN,\n    Filterable: true,\n    Sortable: true\n  })\n];\n\nconst styles: any = {\n  progress: {\n    height: '20px'\n  },\n  search: {\n    margin: '15px 10px 10px 10px',\n    textAlign: 'right'\n  }\n};\n\nclass SampleGridList extends React.Component<IDataGridProps, IDataGridState> {\n  state = { errorMessage: null };\n\n\n  static getDerivedStateFromProps(\n    props: IDataGridProps,\n    state: IDataGridState\n  ) {\n    if (props.error !== state.errorMessage) {\n      return { errorMessage: props.error };\n    }\n    return null;\n  }\n\n  render() {\n    return (\n      <DataSourceContext.Consumer>\n        {({ state }) => (\n          <Paper>\n            <div style={styles.search}>\n              <TextSearchInput />\n            </div>\n            <div style={styles.progress}>\n              {state.isLoading && <LinearProgress />}\n            </div>\n            <Table>\n              <TableBody>\n                <TableRow>\n                  <TableCell>\n                    <GridList cellHeight={180} cols={5}>\n                      {state.data.map(item => (\n                        <GridListTile key={item.OrderID}>\n                          <Card>\n                            <CardContent>\n                              <Typography\n                                gutterBottom={true}\n                                variant='headline'\n                                component='h2'\n                              >\n                                {item.OrderID} - {item.CustomerName}\n                              </Typography>\n                              <Typography component='p'>\n                                {item.ShipperCity}\n                              </Typography>\n                              <Typography component='p'>\n                                {format(item.ShippedDate, 'MMM D YYYY')}\n                              </Typography>\n                            </CardContent>\n                            <CardActions>\n                              <Button size='small' color='primary'>\n                                Learn More\n                              </Button>\n                            </CardActions>\n                          </Card>\n                        </GridListTile>\n                      ))}\n                    </GridList>\n                  </TableCell>\n                </TableRow>\n              </TableBody>\n              <TableFooter>\n                <TableRow>\n                  <Paginator advancePagination={false} />\n                </TableRow>\n              </TableFooter>\n            </Table>\n          </Paper>\n        )}\n      </DataSourceContext.Consumer>\n    );\n  }\n}\n\nexport default withRemoteDataSource(\n  SampleGridList,\n  columns,\n  'https://tubular.azurewebsites.net/api/orders/paged'\n);";
+},{}],"9PWw":[function(require,module,exports) {
+"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var a={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(a[t]=e[t]);return a.default=e,a},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var t=require("@material-ui/core"),r=require("@material-ui/core/styles"),n=e(require("react")),l=a(require("react-highlight")),i=require("../../utils/codeSnipetExamples"),o={code:{fontSize:15},codeTag:{background:"#F8F8FF",padding:"6px"},container:{margin:"0",padding:30,width:"100%"},content:{marginTop:10},image:{marginLeft:"25%",maxWidth:300},paper:{minHeight:"300px",padding:30},paperSmall:{minHeight:"115px",padding:30}},c=function(e){var a=e.classes;return n.createElement(t.Grid,{container:!0,spacing:24,className:a.container},n.createElement(t.Paper,{className:a.paper},n.createElement("img",{src:"..\\..\\static\\tubularRed.png"}),n.createElement(t.Divider,null),n.createElement(t.Typography,{paragraph:!0,variant:"subtitle1",className:a.content},"Tubular React is a set of ReactJS components designed to rapidly build modern web applications. The centerpiece of Tubular is its fully templated grid with lots of features such as server-side pagination, searching text, multi-column sorting, and Filterable, built-in export to CSV or been printed. Another very styled and useful component is the grid list that renders a set of cards with the general information allowing good data visualization and quickly searching, besides it has almost the same grid functionalities."),n.createElement("br",null),n.createElement(t.Typography,{variant:"h4"},"Features"),n.createElement(t.Divider,null),n.createElement(t.Typography,{paragraph:!0,variant:"subtitle1",className:a.content},n.createElement(t.Typography,{variant:"h5"},"The main component is a ",n.createElement("em",null,"grid")," with multiple options:"),n.createElement("ul",null,n.createElement("li",null,"Define a custom layout for columns and cells using render methods."),n.createElement("li",null,"Use a remote or local datasource. Remote datasource use a specific Request and Response format."),n.createElement("li",null,"Sort and filter multiple columns."),n.createElement("li",null,"Free-text search of string columns."),n.createElement("li",null,"Page data. Remote data is paged in the server side."),n.createElement("li",null,"Export data to a CSV file."),n.createElement("li",null,"Print data.")),n.createElement("br",null),"The ",n.createElement("em",null,"grid")," component offers all you can look for an easy grid table and more.",n.createElement("br",null),n.createElement("br",null),n.createElement(t.Typography,{variant:"h5"},n.createElement("em",null,"Grid list")," component allows you to perform some of the same options as:"),n.createElement("ul",null,n.createElement("li",null,"Define a custom layout for columns and cells."),n.createElement("li",null,"Use a remote or local datasource."),n.createElement("li",null,"Free-text search of string columns."),n.createElement("li",null,"Page data. Remote data is paged in the server side.")),n.createElement("br",null),"The ",n.createElement("em",null,"grid list")," component represents another option to show your data.",n.createElement("br",null),n.createElement("em",null,"Grid list")," provide a styled and fancy view over your data, which helps you to quickly find information."),n.createElement("br",null),n.createElement(t.Typography,{variant:"h4",paragraph:!0},"Installation"),n.createElement(t.Divider,null),n.createElement(t.Typography,{variant:"subtitle1",paragraph:!0},"Tubular-React is available as a ",n.createElement("a",{href:"https://www.npmjs.com/package/tubular-react",target:"_blank"},"npm package")),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"Dependencies"),n.createElement(t.Typography,{variant:"body1",paragraph:!0},n.createElement("ul",null,n.createElement("li",null,n.createElement("a",{href:"https://date-fns.org/"},"date-fns - Version: 1.29.0")),n.createElement("li",null,n.createElement("a",{href:"https://material-ui.com/"},"Material-UI - Version: 3.1.0")),n.createElement("li",null,n.createElement("a",{href:"https://reactjs.org/"},"React - Version: 16.5.1")),n.createElement("li",null,n.createElement("a",{href:"https://github.com/unosquare/tubular-common"},"Tubular Common - Version: 1.2.1")))),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"npm"),n.createElement(t.Typography,{variant:"body1",paragraph:!0},"To install and save in your ",n.createElement("code",{className:a.codeTag},"package.json")," dependencies, run:"),n.createElement(l.default,{language:"javascript",className:a.code},"npm install tubular-react --save"),n.createElement("br",null),n.createElement(t.Typography,{variant:"h4",paragraph:!0},"Usage"),n.createElement(t.Divider,null),n.createElement(t.Typography,{variant:"subtitle1",paragraph:!0},"Tubular React is an extension of Material-UI which offers until now a couple of useful components:",n.createElement("br",null),n.createElement("code",{className:a.codeTag},"<DataGrid />")," and",n.createElement("code",{className:a.codeTag},"<GridList />"),".",n.createElement("br",null),"These components supply an organized and nice way to represent your data, you can populate them from a server-side data source or a local data source as an array."),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"Quick Start"),n.createElement(t.Typography,{variant:"body1",paragraph:!0},"Here is a quick example with all basic features to get you started:"),n.createElement(l.default,{language:"javascript",className:a.code},i.basicFeatures),n.createElement(t.Typography,{variant:"h5",paragraph:!0},"Extend Grid Functionalities"),n.createElement(t.Typography,{variant:"subtitle1",paragraph:!0},"You can add functionalities to the `DataGrid`, including extra buttons that can perform an action according to your requirements. You just need to include the IconButton Component from @material-ui and define the icon or button that you need between `DataGrid` tags and specify the action to perform.",n.createElement("br",null),n.createElement(l.default,{language:"javascript",className:a.code},"\n              const MyComponent = () => {\n\n              <DataGrid gridName='Tubular-React'>\n                <IconButton color='default' />\n                  <Brightness7Rounded\n                    onClick={() => alert('Happy codes, have a nice day')}\n                  />\n                </IconButton>\n              </DataGrid>\n\n            };\n              "))))};exports.default=r.withStyles(o)(c);
+},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/","react-highlight":"Xqwm","../../utils/codeSnipetExamples":"Gxw2"}],"vRtn":[function(require,module,exports) {
 "use strict";var e=require("@babel/runtime/helpers/interopRequireDefault");Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var t=e(require("react")),l=e(require("./utils/createSvgIcon")),r=(0,l.default)(t.default.createElement(t.default.Fragment,null,t.default.createElement("path",{fill:"none",d:"M0 0h24v24H0V0z"}),t.default.createElement("path",{d:"M9.4 16.6L4.8 12l4.6-4.6L8 6l-6 6 6 6 1.4-1.4zm5.2 0l4.6-4.6-4.6-4.6L16 6l6 6-6 6-1.4-1.4z"})),"Code");exports.default=r;
 },{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","./utils/createSvgIcon":"Atth"}],"GAz6":[function(require,module,exports) {
 "use strict";var e=require("@babel/runtime/helpers/interopRequireDefault");Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var t=e(require("react")),r=e(require("./utils/createSvgIcon")),l=(0,r.default)(t.default.createElement(t.default.Fragment,null,t.default.createElement("path",{d:"M3 17.25V21h3.75L17.81 9.94l-3.75-3.75L3 17.25zM20.71 7.04c.39-.39.39-1.02 0-1.41l-2.34-2.34a.9959.9959 0 0 0-1.41 0l-1.83 1.83 3.75 3.75 1.83-1.83z"}),t.default.createElement("path",{fill:"none",d:"M0 0h24v24H0z"})),"Create");exports.default=l;
-},{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","./utils/createSvgIcon":"Atth"}],"07Mr":[function(require,module,exports) {
+},{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","./utils/createSvgIcon":"Atth"}],"W7b0":[function(require,module,exports) {
+
+},{}],"07Mr":[function(require,module,exports) {
 "use strict";var e=require("@babel/runtime/helpers/interopRequireDefault");Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var t=e(require("react")),r=e(require("./utils/createSvgIcon")),a=(0,r.default)(t.default.createElement(t.default.Fragment,null,t.default.createElement("path",{fill:"none",d:"M0 0h24v24H0z"}),t.default.createElement("path",{d:"M1 21h22L12 2 1 21zm12-3h-2v-2h2v2zm0-4h-2v-4h2v4z"})),"Warning");exports.default=a;
 },{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","./utils/createSvgIcon":"Atth"}],"AtJo":[function(require,module,exports) {
 "use strict";var t=this&&this.__importStar||function(t){if(t&&t.__esModule)return t;var e={};if(null!=t)for(var r in t)Object.hasOwnProperty.call(t,r)&&(e[r]=t[r]);return e.default=t,e};exports.__esModule=!0;var e=t(require("react"));exports.DataSourceContext=e.createContext({actions:null,state:null});
@@ -1711,7 +1691,7 @@ var r=require("../add_years/index.js");function e(e,u){var a=Number(u);return r(
 },{"../add_years/index.js":"WGS0"}],"cWQX":[function(require,module,exports) {
 module.exports={addDays:require("./add_days/index.js"),addHours:require("./add_hours/index.js"),addISOYears:require("./add_iso_years/index.js"),addMilliseconds:require("./add_milliseconds/index.js"),addMinutes:require("./add_minutes/index.js"),addMonths:require("./add_months/index.js"),addQuarters:require("./add_quarters/index.js"),addSeconds:require("./add_seconds/index.js"),addWeeks:require("./add_weeks/index.js"),addYears:require("./add_years/index.js"),areRangesOverlapping:require("./are_ranges_overlapping/index.js"),closestIndexTo:require("./closest_index_to/index.js"),closestTo:require("./closest_to/index.js"),compareAsc:require("./compare_asc/index.js"),compareDesc:require("./compare_desc/index.js"),differenceInCalendarDays:require("./difference_in_calendar_days/index.js"),differenceInCalendarISOWeeks:require("./difference_in_calendar_iso_weeks/index.js"),differenceInCalendarISOYears:require("./difference_in_calendar_iso_years/index.js"),differenceInCalendarMonths:require("./difference_in_calendar_months/index.js"),differenceInCalendarQuarters:require("./difference_in_calendar_quarters/index.js"),differenceInCalendarWeeks:require("./difference_in_calendar_weeks/index.js"),differenceInCalendarYears:require("./difference_in_calendar_years/index.js"),differenceInDays:require("./difference_in_days/index.js"),differenceInHours:require("./difference_in_hours/index.js"),differenceInISOYears:require("./difference_in_iso_years/index.js"),differenceInMilliseconds:require("./difference_in_milliseconds/index.js"),differenceInMinutes:require("./difference_in_minutes/index.js"),differenceInMonths:require("./difference_in_months/index.js"),differenceInQuarters:require("./difference_in_quarters/index.js"),differenceInSeconds:require("./difference_in_seconds/index.js"),differenceInWeeks:require("./difference_in_weeks/index.js"),differenceInYears:require("./difference_in_years/index.js"),distanceInWords:require("./distance_in_words/index.js"),distanceInWordsStrict:require("./distance_in_words_strict/index.js"),distanceInWordsToNow:require("./distance_in_words_to_now/index.js"),eachDay:require("./each_day/index.js"),endOfDay:require("./end_of_day/index.js"),endOfHour:require("./end_of_hour/index.js"),endOfISOWeek:require("./end_of_iso_week/index.js"),endOfISOYear:require("./end_of_iso_year/index.js"),endOfMinute:require("./end_of_minute/index.js"),endOfMonth:require("./end_of_month/index.js"),endOfQuarter:require("./end_of_quarter/index.js"),endOfSecond:require("./end_of_second/index.js"),endOfToday:require("./end_of_today/index.js"),endOfTomorrow:require("./end_of_tomorrow/index.js"),endOfWeek:require("./end_of_week/index.js"),endOfYear:require("./end_of_year/index.js"),endOfYesterday:require("./end_of_yesterday/index.js"),format:require("./format/index.js"),getDate:require("./get_date/index.js"),getDay:require("./get_day/index.js"),getDayOfYear:require("./get_day_of_year/index.js"),getDaysInMonth:require("./get_days_in_month/index.js"),getDaysInYear:require("./get_days_in_year/index.js"),getHours:require("./get_hours/index.js"),getISODay:require("./get_iso_day/index.js"),getISOWeek:require("./get_iso_week/index.js"),getISOWeeksInYear:require("./get_iso_weeks_in_year/index.js"),getISOYear:require("./get_iso_year/index.js"),getMilliseconds:require("./get_milliseconds/index.js"),getMinutes:require("./get_minutes/index.js"),getMonth:require("./get_month/index.js"),getOverlappingDaysInRanges:require("./get_overlapping_days_in_ranges/index.js"),getQuarter:require("./get_quarter/index.js"),getSeconds:require("./get_seconds/index.js"),getTime:require("./get_time/index.js"),getYear:require("./get_year/index.js"),isAfter:require("./is_after/index.js"),isBefore:require("./is_before/index.js"),isDate:require("./is_date/index.js"),isEqual:require("./is_equal/index.js"),isFirstDayOfMonth:require("./is_first_day_of_month/index.js"),isFriday:require("./is_friday/index.js"),isFuture:require("./is_future/index.js"),isLastDayOfMonth:require("./is_last_day_of_month/index.js"),isLeapYear:require("./is_leap_year/index.js"),isMonday:require("./is_monday/index.js"),isPast:require("./is_past/index.js"),isSameDay:require("./is_same_day/index.js"),isSameHour:require("./is_same_hour/index.js"),isSameISOWeek:require("./is_same_iso_week/index.js"),isSameISOYear:require("./is_same_iso_year/index.js"),isSameMinute:require("./is_same_minute/index.js"),isSameMonth:require("./is_same_month/index.js"),isSameQuarter:require("./is_same_quarter/index.js"),isSameSecond:require("./is_same_second/index.js"),isSameWeek:require("./is_same_week/index.js"),isSameYear:require("./is_same_year/index.js"),isSaturday:require("./is_saturday/index.js"),isSunday:require("./is_sunday/index.js"),isThisHour:require("./is_this_hour/index.js"),isThisISOWeek:require("./is_this_iso_week/index.js"),isThisISOYear:require("./is_this_iso_year/index.js"),isThisMinute:require("./is_this_minute/index.js"),isThisMonth:require("./is_this_month/index.js"),isThisQuarter:require("./is_this_quarter/index.js"),isThisSecond:require("./is_this_second/index.js"),isThisWeek:require("./is_this_week/index.js"),isThisYear:require("./is_this_year/index.js"),isThursday:require("./is_thursday/index.js"),isToday:require("./is_today/index.js"),isTomorrow:require("./is_tomorrow/index.js"),isTuesday:require("./is_tuesday/index.js"),isValid:require("./is_valid/index.js"),isWednesday:require("./is_wednesday/index.js"),isWeekend:require("./is_weekend/index.js"),isWithinRange:require("./is_within_range/index.js"),isYesterday:require("./is_yesterday/index.js"),lastDayOfISOWeek:require("./last_day_of_iso_week/index.js"),lastDayOfISOYear:require("./last_day_of_iso_year/index.js"),lastDayOfMonth:require("./last_day_of_month/index.js"),lastDayOfQuarter:require("./last_day_of_quarter/index.js"),lastDayOfWeek:require("./last_day_of_week/index.js"),lastDayOfYear:require("./last_day_of_year/index.js"),max:require("./max/index.js"),min:require("./min/index.js"),parse:require("./parse/index.js"),setDate:require("./set_date/index.js"),setDay:require("./set_day/index.js"),setDayOfYear:require("./set_day_of_year/index.js"),setHours:require("./set_hours/index.js"),setISODay:require("./set_iso_day/index.js"),setISOWeek:require("./set_iso_week/index.js"),setISOYear:require("./set_iso_year/index.js"),setMilliseconds:require("./set_milliseconds/index.js"),setMinutes:require("./set_minutes/index.js"),setMonth:require("./set_month/index.js"),setQuarter:require("./set_quarter/index.js"),setSeconds:require("./set_seconds/index.js"),setYear:require("./set_year/index.js"),startOfDay:require("./start_of_day/index.js"),startOfHour:require("./start_of_hour/index.js"),startOfISOWeek:require("./start_of_iso_week/index.js"),startOfISOYear:require("./start_of_iso_year/index.js"),startOfMinute:require("./start_of_minute/index.js"),startOfMonth:require("./start_of_month/index.js"),startOfQuarter:require("./start_of_quarter/index.js"),startOfSecond:require("./start_of_second/index.js"),startOfToday:require("./start_of_today/index.js"),startOfTomorrow:require("./start_of_tomorrow/index.js"),startOfWeek:require("./start_of_week/index.js"),startOfYear:require("./start_of_year/index.js"),startOfYesterday:require("./start_of_yesterday/index.js"),subDays:require("./sub_days/index.js"),subHours:require("./sub_hours/index.js"),subISOYears:require("./sub_iso_years/index.js"),subMilliseconds:require("./sub_milliseconds/index.js"),subMinutes:require("./sub_minutes/index.js"),subMonths:require("./sub_months/index.js"),subQuarters:require("./sub_quarters/index.js"),subSeconds:require("./sub_seconds/index.js"),subWeeks:require("./sub_weeks/index.js"),subYears:require("./sub_years/index.js")};
 },{"./add_days/index.js":"C8Jj","./add_hours/index.js":"cFoC","./add_iso_years/index.js":"bhh0","./add_milliseconds/index.js":"JXAh","./add_minutes/index.js":"+vZp","./add_months/index.js":"yVyG","./add_quarters/index.js":"H1DN","./add_seconds/index.js":"Gepr","./add_weeks/index.js":"hDyn","./add_years/index.js":"WGS0","./are_ranges_overlapping/index.js":"8tx8","./closest_index_to/index.js":"K0rp","./closest_to/index.js":"xeDi","./compare_asc/index.js":"+NJC","./compare_desc/index.js":"pfSt","./difference_in_calendar_days/index.js":"7dmj","./difference_in_calendar_iso_weeks/index.js":"OCNd","./difference_in_calendar_iso_years/index.js":"KVP/","./difference_in_calendar_months/index.js":"mgfB","./difference_in_calendar_quarters/index.js":"8/M/","./difference_in_calendar_weeks/index.js":"pfXU","./difference_in_calendar_years/index.js":"3TH0","./difference_in_days/index.js":"F0zg","./difference_in_hours/index.js":"lvwL","./difference_in_iso_years/index.js":"32aN","./difference_in_milliseconds/index.js":"KKv9","./difference_in_minutes/index.js":"F5+8","./difference_in_months/index.js":"5KzI","./difference_in_quarters/index.js":"A399","./difference_in_seconds/index.js":"Sjwe","./difference_in_weeks/index.js":"E2bU","./difference_in_years/index.js":"3SSW","./distance_in_words/index.js":"HLwM","./distance_in_words_strict/index.js":"WrFA","./distance_in_words_to_now/index.js":"jUDc","./each_day/index.js":"jayk","./end_of_day/index.js":"hjyj","./end_of_hour/index.js":"Ytlb","./end_of_iso_week/index.js":"1Xfk","./end_of_iso_year/index.js":"p9Kg","./end_of_minute/index.js":"y1Cm","./end_of_month/index.js":"3mb3","./end_of_quarter/index.js":"CR29","./end_of_second/index.js":"W3u2","./end_of_today/index.js":"aAfn","./end_of_tomorrow/index.js":"nuKj","./end_of_week/index.js":"wsAK","./end_of_year/index.js":"zeTk","./end_of_yesterday/index.js":"37Up","./format/index.js":"Vy4H","./get_date/index.js":"Qptf","./get_day/index.js":"3CsB","./get_day_of_year/index.js":"5S8c","./get_days_in_month/index.js":"AF7/","./get_days_in_year/index.js":"HHnJ","./get_hours/index.js":"4vgY","./get_iso_day/index.js":"+P0g","./get_iso_week/index.js":"eL1o","./get_iso_weeks_in_year/index.js":"f/Mr","./get_iso_year/index.js":"7L22","./get_milliseconds/index.js":"iJrm","./get_minutes/index.js":"YRfl","./get_month/index.js":"V3Bo","./get_overlapping_days_in_ranges/index.js":"mYqR","./get_quarter/index.js":"TxW9","./get_seconds/index.js":"gFWC","./get_time/index.js":"+HzN","./get_year/index.js":"bxn3","./is_after/index.js":"+i84","./is_before/index.js":"dR3O","./is_date/index.js":"CFBi","./is_equal/index.js":"Gp5+","./is_first_day_of_month/index.js":"NYcv","./is_friday/index.js":"CMn2","./is_future/index.js":"ngA7","./is_last_day_of_month/index.js":"6qwu","./is_leap_year/index.js":"SThF","./is_monday/index.js":"aaZC","./is_past/index.js":"5OHs","./is_same_day/index.js":"3Ehu","./is_same_hour/index.js":"6NfB","./is_same_iso_week/index.js":"s6JR","./is_same_iso_year/index.js":"Dwal","./is_same_minute/index.js":"Wmy4","./is_same_month/index.js":"9dRT","./is_same_quarter/index.js":"lXxI","./is_same_second/index.js":"0pSd","./is_same_week/index.js":"H3Kb","./is_same_year/index.js":"Qshd","./is_saturday/index.js":"zGeD","./is_sunday/index.js":"ct5/","./is_this_hour/index.js":"gMKA","./is_this_iso_week/index.js":"FptX","./is_this_iso_year/index.js":"+hsk","./is_this_minute/index.js":"dBgs","./is_this_month/index.js":"gjob","./is_this_quarter/index.js":"xJQv","./is_this_second/index.js":"zCzs","./is_this_week/index.js":"V1rv","./is_this_year/index.js":"2q5b","./is_thursday/index.js":"fsMm","./is_today/index.js":"VNQV","./is_tomorrow/index.js":"28nI","./is_tuesday/index.js":"fK+1","./is_valid/index.js":"x3VB","./is_wednesday/index.js":"G+9b","./is_weekend/index.js":"jxMk","./is_within_range/index.js":"g8eB","./is_yesterday/index.js":"NOli","./last_day_of_iso_week/index.js":"xwmC","./last_day_of_iso_year/index.js":"bzJd","./last_day_of_month/index.js":"Zd9R","./last_day_of_quarter/index.js":"EffO","./last_day_of_week/index.js":"8M3P","./last_day_of_year/index.js":"ps1A","./max/index.js":"TgIO","./min/index.js":"yan+","./parse/index.js":"ajCz","./set_date/index.js":"B65t","./set_day/index.js":"ADPD","./set_day_of_year/index.js":"oUK2","./set_hours/index.js":"LmDY","./set_iso_day/index.js":"geG1","./set_iso_week/index.js":"WWMt","./set_iso_year/index.js":"C4o3","./set_milliseconds/index.js":"ur5H","./set_minutes/index.js":"0gvM","./set_month/index.js":"FE9H","./set_quarter/index.js":"ES+P","./set_seconds/index.js":"9lKv","./set_year/index.js":"xEl2","./start_of_day/index.js":"GCdn","./start_of_hour/index.js":"8fe4","./start_of_iso_week/index.js":"Rrov","./start_of_iso_year/index.js":"z+C4","./start_of_minute/index.js":"Y6PB","./start_of_month/index.js":"D/sO","./start_of_quarter/index.js":"EcBT","./start_of_second/index.js":"yE+e","./start_of_today/index.js":"sUjp","./start_of_tomorrow/index.js":"iR7a","./start_of_week/index.js":"bGge","./start_of_year/index.js":"SInJ","./start_of_yesterday/index.js":"RGBM","./sub_days/index.js":"02fE","./sub_hours/index.js":"oWR9","./sub_iso_years/index.js":"jrl+","./sub_milliseconds/index.js":"4ip+","./sub_minutes/index.js":"wdQF","./sub_months/index.js":"vzmz","./sub_quarters/index.js":"3L28","./sub_seconds/index.js":"wwfT","./sub_weeks/index.js":"J4St","./sub_years/index.js":"18SO"}],"y1lj":[function(require,module,exports) {
-"use strict";var e=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}},r=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var r={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(r[t]=e[t]);return r.default=e,r};exports.__esModule=!0;var t=e(require("@material-ui/core/TableCell")),a=e(require("@material-ui/icons/CheckBox")),n=e(require("@material-ui/icons/CheckBoxOutlineBlank")),u=require("date-fns"),l=r(require("react")),i=require("tubular-common");exports.renderCellContent=function(e,r){switch(e.DataType){case i.ColumnDataType.NUMERIC:return r[e.Name]||0;case i.ColumnDataType.DATE:return r[e.Name]&&u.getYear(u.parse(r[e.Name]))>0?u.format(r[e.Name],"M/D/YYYY"):"";case i.ColumnDataType.DATE_TIME:case i.ColumnDataType.DATE_TIME_UTC:return r[e.Name]&&u.getYear(u.parse(r[e.Name]))>0?u.format(r[e.Name],"M/D/YYYY h:mm A"):"";case i.ColumnDataType.BOOLEAN:return!0===r[e.Name]?l.createElement(a.default,null):l.createElement(n.default,null);default:return r[e.Name]}},exports.renderCells=function(e,r){return e.filter(function(e){return e.Visible}).map(function(e,a){return l.createElement(t.default,{key:a,padding:""===e.Label?"none":"default"},exports.renderCellContent(e,r))})};
+"use strict";var e=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}},r=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var r={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(r[t]=e[t]);return r.default=e,r};exports.__esModule=!0;var t=e(require("@material-ui/core/TableCell")),a=e(require("@material-ui/icons/CheckBox")),n=e(require("@material-ui/icons/CheckBoxOutlineBlank")),u=require("date-fns"),l=r(require("react")),i=require("tubular-common");exports.renderCellContent=function(e,r){switch(e.DataType){case i.ColumnDataType.NUMERIC:return r[e.Name]||0;case i.ColumnDataType.DATE:return r[e.Name]&&u.getYear(u.parse(r[e.Name]))>0?u.format(r[e.Name],"M/D/YYYY"):"";case i.ColumnDataType.DATE_TIME:case i.ColumnDataType.DATE_TIME_UTC:return r[e.Name]&&u.getYear(u.parse(r[e.Name]))>0?u.format(r[e.Name],"M/D/YYYY h:mm A"):"";case i.ColumnDataType.BOOLEAN:return!0===r[e.Name]?l.createElement(a.default,null):l.createElement(n.default,null);default:return r[e.Name]}},exports.renderCells=function(e,r){return e.filter(function(e){return e.Visible}).map(function(e){return l.createElement(t.default,{key:e.Name,padding:""===e.Label?"none":"default"},exports.renderCellContent(e,r))})};
 },{"@material-ui/core/TableCell":"b+QQ","@material-ui/icons/CheckBox":"ou/W","@material-ui/icons/CheckBoxOutlineBlank":"lqdk","date-fns":"cWQX","react":"1n8/","tubular-common":"QwmZ"}],"UeD7":[function(require,module,exports) {
 "use strict";exports.__esModule=!0;var t=function(){function t(t,e){this.request=t,this.handler=e}return t.prototype.fetch=function(t){return this.handler(this.request,t)},t}();exports.default=t;
 },{}],"xhdv":[function(require,module,exports) {
@@ -1733,7 +1713,7 @@ module.exports={addDays:require("./add_days/index.js"),addHours:require("./add_h
 },{"@babel/runtime/helpers/interopRequireDefault":"SpGf","react":"1n8/","./utils/createSvgIcon":"Atth"}],"0kQ0":[function(require,module,exports) {
 "use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var a in e)Object.hasOwnProperty.call(e,a)&&(t[a]=e[a]);return t.default=e,t},t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var a=require("date-fns"),r=e(require("react")),u=t(require("@material-ui/core/MenuItem")),l=t(require("@material-ui/core/TextField")),n=require("tubular-common"),o={marginLeft:"10%",marginTop:"5%",width:"80%"},i=[{Value:"true",Title:"True"},{Value:"false",Title:"False"}],m={boolean:"text",date:"date",datetime:"datetime-local",datetimeutc:"datetime-local",numeric:"number",string:"text"},s=function(e,t,r,u){switch(e){case n.ColumnDataType.DATE:return r?a.format(r,"YYYY-MM-DD"):(u(a.format(new Date,"YYYY-MM-DD")),"");case n.ColumnDataType.DATE_TIME:case n.ColumnDataType.DATE_TIME_UTC:return r?a.format(r,"YYYY-MM-DD[T]HH:mm"):(u(a.format(new Date,"YYYY-MM-DD[T]HH:mm")),"");case n.ColumnDataType.BOOLEAN:return t===n.CompareOperators.NONE?"":"boolean"==typeof r?!0===r?"true":"false":r;default:return t===n.CompareOperators.NONE?"":r||""}},p=function(e){var t=e.column,a=e.handleTextFieldChange,p=e.isPrimary,c=s(t.DataType,t.Operator,p?t.Filter.Text:t.Filter.Argument[0],a)||"",T=!!p&&t.Filter.Operator===n.CompareOperators.NONE,f=p?t.Filter.Operator!==n.CompareOperators.BETWEEN?"Value":"First Value":"Second Value";return r.createElement(l.default,{select:t.DataType===n.ColumnDataType.BOOLEAN,style:o,id:t.Name,disabled:T,value:c,label:f,type:m[t.DataType],onChange:function(e){return a(e.target.value)}},t.DataType===n.ColumnDataType.BOOLEAN&&i.map(function(e){return r.createElement(u.default,{key:e.Value,value:e.Value},e.Title)}))};exports.default=p;
 },{"date-fns":"cWQX","react":"1n8/","@material-ui/core/MenuItem":"MsWr","@material-ui/core/TextField":"zR0T","tubular-common":"QwmZ"}],"zLYP":[function(require,module,exports) {
-"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var r in e)Object.hasOwnProperty.call(e,r)&&(t[r]=e[r]);return t.default=e,t},t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var r=e(require("react")),u=t(require("@material-ui/core/Input")),a=t(require("@material-ui/core/MenuItem")),n=t(require("@material-ui/core/Select")),i=require("tubular-common"),l=require("../DataSource"),o={marginLeft:"10%",width:"80%"},c=function(e){return function(t){var r=t.target;return e({Operator:r.value})}},m=function(e){return e&&""!==e?e:"None"},f=function(){return r.createElement(l.DataSourceContext.Consumer,null,function(e){var t=e.state,l=e.actions;return r.createElement(n.default,{style:o,value:m(t.activeColumn.Filter),onChange:c(l.handleFilterChange),input:r.createElement(u.default,{name:t.activeColumn.Name})},i.ColumnModel.getOperators(t.activeColumn).map(function(e){return r.createElement(a.default,{key:e.Value,value:e.Value},e.Title)}))})};exports.default=f;
+"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var r in e)Object.hasOwnProperty.call(e,r)&&(t[r]=e[r]);return t.default=e,t},t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var r=e(require("react")),u=t(require("@material-ui/core/Input")),a=t(require("@material-ui/core/MenuItem")),n=t(require("@material-ui/core/Select")),i=require("tubular-common"),l=require("../DataSource"),o={marginLeft:"10%",width:"80%"},c=function(e){return function(t){var r=t.target;return e({Operator:r.value})}},m=function(e){return e&&""!==e?e:"None"},f=function(){return r.createElement(l.DataSourceContext.Consumer,null,function(e){var t=e.state,l=e.actions;return r.createElement(n.default,{style:o,value:m(t.activeColumn.Filter.Operator),onChange:c(l.handleFilterChange),input:r.createElement(u.default,{name:t.activeColumn.Name})},i.ColumnModel.getOperators(t.activeColumn).map(function(e){return r.createElement(a.default,{key:e.Value,value:e.Value},e.Title)}))})};exports.default=f;
 },{"react":"1n8/","@material-ui/core/Input":"V7QS","@material-ui/core/MenuItem":"MsWr","@material-ui/core/Select":"bsjm","tubular-common":"QwmZ","../DataSource":"aB8x"}],"BxX0":[function(require,module,exports) {
 "use strict";var e=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}},r=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var r={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(r[t]=e[t]);return r.default=e,r};exports.__esModule=!0;var t=e(require("@material-ui/core/Button")),a=e(require("@material-ui/core/Card")),n=e(require("@material-ui/core/CardActions")),l=e(require("@material-ui/core/CardContent")),u=e(require("@material-ui/core/Popover")),o=r(require("react")),i=require("../DataSource"),c=e(require("./DialogInput")),m=e(require("./OperatorsDropdown")),s=require("tubular-common"),d=function(e){var r=e.Filter.Text,t=e.Filter.Argument[0];return e.DataType===s.ColumnDataType.NUMERIC?(r=parseFloat(r),t=parseFloat(t)):e.DataType===s.ColumnDataType.BOOLEAN&&(r="true"===r,t=""),{Argument:[t],HasFilter:!0,Text:r}},f={Argument:[""],HasFilter:!1,Operator:s.CompareOperators.NONE,Text:""},p=function(e){return function(){return e(f)}},C=function(e){return function(r){return e({Text:r})}},h=function(e){return function(r){return e({Argument:[r]})}},E=function(e,r){return function(){return e(d(r))}},F=function(){return o.createElement(i.DataSourceContext.Consumer,null,function(e){var r=e.state,i=e.actions;return o.createElement(u.default,{open:Boolean(r.anchorFilter),onClose:p(i.setFilter),anchorEl:r.anchorFilter,anchorOrigin:{horizontal:"center",vertical:"bottom"},transformOrigin:{horizontal:"center",vertical:"top"}},o.createElement(a.default,null,o.createElement(l.default,null,o.createElement(m.default,null),o.createElement(c.default,{column:r.activeColumn,isPrimary:!0,handleTextFieldChange:C(i.handleFilterChange)}),r.activeColumn.Filter.Operator===s.CompareOperators.BETWEEN&&o.createElement(c.default,{column:r.activeColumn,isPrimary:!1,handleTextFieldChange:h(i.handleFilterChange)})),o.createElement(n.default,null,o.createElement(t.default,{size:"medium",color:"secondary",onClick:p(i.setFilter)},"Clear"),o.createElement(t.default,{size:"medium",color:"primary",onClick:E(i.setFilter,r.activeColumn),disabled:r.activeColumn.Filter.Operator===s.CompareOperators.NONE},"Apply"))))})};exports.default=F;
 },{"@material-ui/core/Button":"Pl3a","@material-ui/core/Card":"wK/i","@material-ui/core/CardActions":"OF+c","@material-ui/core/CardContent":"Ij76","@material-ui/core/Popover":"Dxrq","react":"1n8/","../DataSource":"aB8x","./DialogInput":"0kQ0","./OperatorsDropdown":"zLYP","tubular-common":"QwmZ"}],"CcII":[function(require,module,exports) {
@@ -1790,8 +1770,16 @@ module.exports={addDays:require("./add_days/index.js"),addHours:require("./add_h
 "use strict";var e=this&&this.__extends||function(){var e=function(t,r){return(e=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var r in t)t.hasOwnProperty(r)&&(e[r]=t[r])})(t,r)};return function(t,r){function l(){this.constructor=t}e(t,r),t.prototype=null===r?Object.create(r):(l.prototype=r.prototype,new l)}}(),t=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var r in e)Object.hasOwnProperty.call(e,r)&&(t[r]=e[r]);return t.default=e,t},r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var l=t(require("react")),a=require("@material-ui/core"),n=r(require("@material-ui/icons/CheckBox")),o=r(require("@material-ui/icons/CheckBoxOutlineBlank")),i=require("date-fns"),u=t(require("../../src")),s=r(require("../utils/columns")),c=function(t){function r(){var e=null!==t&&t.apply(this,arguments)||this;return e.state={errorMessage:null},e}return e(r,t),r.prototype.componentWillReceiveProps=function(e){this.setState({errorMessage:e.error})},r.prototype.render=function(){var e=this.state.errorMessage;return l.createElement("div",{className:"root"},e&&l.createElement(a.Snackbar,{anchorOrigin:{vertical:"top",horizontal:"center"},style:{paddingTop:"10px"},open:!0,ContentProps:{"aria-describedby":"message-id"},message:l.createElement("span",{id:"message-id"},e)}),l.createElement(u.default,{gridName:"Tubular-React",bodyRenderer:function(e){return l.createElement(a.TableRow,{hover:!0,key:e},l.createElement(a.TableCell,{padding:"default"},e.OrderID),l.createElement(a.TableCell,{padding:"default"},e.CustomerName),l.createElement(a.TableCell,{padding:"default"},i.format(e.ShippedDate,"MMMM Do YYYY, h:mm:ss a")),l.createElement(a.TableCell,{padding:"default"},e.ShipperCity),l.createElement(a.TableCell,{padding:"default",numeric:!0},e.Amount||0),l.createElement(a.TableCell,{padding:"default"},e.IsShipped?l.createElement(n.default,null):l.createElement(o.default,null)))},toolbarOptions:new u.ToolbarOptions,footerRenderer:function(e){return l.createElement(a.TableRow,null,l.createElement(a.TableCell,null,"Total: "),l.createElement(a.TableCell,null,e&&e.CustomerName),l.createElement(a.TableCell,null),l.createElement(a.TableCell,null),l.createElement(a.TableCell,null))}}))},r}(l.Component);exports.default=u.withRemoteDataSource(c,s.default,"https://tubular.azurewebsites.net/api/orders/paged");
 },{"react":"1n8/","@material-ui/core":"J2ra","@material-ui/icons/CheckBox":"ou/W","@material-ui/icons/CheckBoxOutlineBlank":"lqdk","date-fns":"cWQX","../../src":"+fUd","../utils/columns":"jRiX"}],"puvW":[function(require,module,exports) {
 "use strict";var e=this&&this.__extends||function(){var e=function(t,r){return(e=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var r in t)t.hasOwnProperty(r)&&(e[r]=t[r])})(t,r)};return function(t,r){function n(){this.constructor=t}e(t,r),t.prototype=null===r?Object.create(r):(n.prototype=r.prototype,new n)}}(),t=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var r in e)Object.hasOwnProperty.call(e,r)&&(t[r]=e[r]);return t.default=e,t},r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var n=require("@material-ui/core"),a=require("@material-ui/core"),l=require("@material-ui/core"),o=require("@material-ui/core"),i=require("@material-ui/core"),u=require("date-fns"),c=t(require("react")),s=r(require("../utils/columns")),p=require("../../src"),m={progress:{height:"20px"},search:{margin:"15px 10px 10px 10px",textAlign:"right"}},d=function(t){function r(){var e=null!==t&&t.apply(this,arguments)||this;return e.state={errorMessage:null},e}return e(r,t),r.getDerivedStateFromProps=function(e,t){return e.error!==t.errorMessage?{errorMessage:e.error}:null},r.prototype.render=function(){return c.createElement(p.DataSourceContext.Consumer,null,function(e){var t=e.state;return c.createElement(n.Paper,null,c.createElement("div",{style:m.search},c.createElement(p.TextSearchInput,null)),c.createElement("div",{style:m.progress},t.isLoading&&c.createElement(n.LinearProgress,null)),c.createElement(i.Table,null,c.createElement(i.TableBody,null,c.createElement(i.TableRow,null,c.createElement(i.TableCell,null,c.createElement(l.GridList,{cellHeight:180,cols:5},t.data.map(function(e){return c.createElement(l.GridListTile,{key:e.OrderID},c.createElement(a.Card,null,c.createElement(a.CardContent,null,c.createElement(l.Typography,{gutterBottom:!0,variant:"h5",component:"h2"},e.OrderID," - ",e.CustomerName),c.createElement(l.Typography,{component:"p"},e.ShipperCity),c.createElement(l.Typography,{component:"p"},u.format(e.ShippedDate,"MMM D YYYY"))),c.createElement(a.CardActions,null,c.createElement(o.Button,{size:"small",color:"primary"},"Learn More"))))}))))),c.createElement(i.TableFooter,null,c.createElement(i.TableRow,null,c.createElement(p.Paginator,{advancePagination:!1})))))})},r}(c.Component);exports.default=p.withRemoteDataSource(d,s.default,"https://tubular.azurewebsites.net/api/orders/paged");
-},{"@material-ui/core":"J2ra","date-fns":"cWQX","react":"1n8/","../utils/columns":"jRiX","../../src":"+fUd"}],"rxup":[function(require,module,exports) {
-"use strict";var e=this&&this.__extends||function(){var e=function(t,a){return(e=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var a in t)t.hasOwnProperty(a)&&(e[a]=t[a])})(t,a)};return function(t,a){function n(){this.constructor=t}e(t,a),t.prototype=null===a?Object.create(a):(n.prototype=a.prototype,new n)}}(),t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}},a=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var a in e)Object.hasOwnProperty.call(e,a)&&(t[a]=e[a]);return t.default=e,t};exports.__esModule=!0;var n=require("@material-ui/core"),r=require("@material-ui/core"),l=require("@material-ui/core"),o=t(require("@material-ui/icons/Code")),i=t(require("@material-ui/icons/Create")),c=t(require("react-highlight")),s=t(require("../components/Github")),u=a(require("react"));require("highlight.js/styles/an-old-hope.css");var m=t(require("../../srcdocs/components/BasicFeatures")),p=t(require("../../srcdocs/components/SampleButtonsFeatures")),d=t(require("../../srcdocs/components/SampleFeatures")),h=t(require("../../srcdocs/components/SampleGridList")),E=require("../utils/codeSnipetExamples"),f={code:{fontSize:15},container:{margin:"0",padding:30,width:"100%"},content:{marginTop:10},icon:{height:24,viewBox:"0 0 24 24",width:24},paper:{padding:20}},g={textAlign:"right"};exports.default=n.withStyles(f)(function(t){function a(){var e=null!==t&&t.apply(this,arguments)||this;return e.state={openBasic:!1,openGridList:!1,openSample:!1,showCode:!1,value:"one"},e.handleClickBasic=function(){e.setState({openBasic:!e.state.openBasic})},e.handleClickSample=function(){e.setState({openSample:!e.state.openSample})},e.handleClickGrid=function(){e.setState({openGridList:!e.state.openGridList})},e}return e(a,t),a.prototype.render=function(){var e=this.props.classes;return u.createElement("div",null,u.createElement(n.Grid,{container:!0,spacing:24,className:e.container},u.createElement(n.Grid,{item:!0,xs:12,className:e.paper},u.createElement(n.Paper,{className:e.paper},u.createElement(n.Typography,{variant:"h4"},"Samples"),u.createElement(n.Divider,null),u.createElement("br",null),u.createElement(n.Grid,{item:!0,xs:12,className:e.paper},u.createElement(n.Typography,{variant:"h5",gutterBottom:!0},"Basic grid features"),u.createElement(n.Typography,{variant:"h6"},"Very basic grid sample"),u.createElement(n.Grid,{item:!0,xs:12},u.createElement("div",{style:g},u.createElement(l.Tooltip,{title:"Tubular GitHub Repo"},u.createElement(l.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",color:"inherit"},u.createElement(s.default,null))),u.createElement(l.Tooltip,{title:this.state.openBasic?"Hide Code":"View Code"},u.createElement(l.IconButton,{onClick:this.handleClickBasic},u.createElement(o.default,null))),u.createElement(l.Tooltip,{title:"Edit in CodeSandbox"},u.createElement(l.IconButton,{component:"a",href:"https://codesandbox.io/s/lp594wqr1q",target:"_blank",color:"inherit"},u.createElement(i.default,null)))),u.createElement(r.Collapse,{in:this.state.openBasic,timeout:"auto"},u.createElement(n.Paper,null,u.createElement(c.default,{language:"javascript",className:e.code},E.basicFeatures))),u.createElement(m.default,null))),u.createElement(n.Divider,null),u.createElement(n.Grid,{item:!0,xs:12,className:e.paper},u.createElement(n.Typography,{variant:"h5",gutterBottom:!0},"Full grid features"),u.createElement(n.Typography,{variant:"h6"},"The grid can be extended to include features like sorting, Filterable, free-text searches, move across the pages and change the number of rows per page Although easily you can print or export the current view or entire dataset to CSV using client-side only."),u.createElement(n.Grid,{item:!0,xs:12},u.createElement("div",{style:g},u.createElement(l.Tooltip,{title:"Tubular GitHub Repo"},u.createElement(l.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",color:"inherit"},u.createElement(s.default,null))),u.createElement(l.Tooltip,{title:this.state.openSample?"Hide Code":"View Code"},u.createElement(l.IconButton,{onClick:this.handleClickSample},u.createElement(o.default,null))),u.createElement(l.Tooltip,{title:"Edit in CodeSandbox"},u.createElement(l.IconButton,{component:"a",href:"https://codesandbox.io/s/434rlkr2y9",target:"_blank",color:"inherit"},u.createElement(i.default,null)))),u.createElement(r.Collapse,{in:this.state.openSample,timeout:"auto"},u.createElement(n.Paper,null,u.createElement(c.default,{language:"javascript",className:e.code},E.simpleFeatures))),u.createElement(d.default,null))),u.createElement(n.Divider,null),u.createElement(n.Grid,{item:!0,xs:12,className:e.paper},u.createElement(n.Typography,{variant:"h5",gutterBottom:!0},"How to include functionality buttons"),u.createElement(n.Typography,{variant:"h6"},"You can add functionalities to the DataGrid including extra buttons that can perform an action according to your requirements. Just need include an IconButton Component from @material-ui and define the icon or button that you need between DataGrid tags and specify the action to perform."),u.createElement(n.Grid,{item:!0,xs:12},u.createElement("div",{style:g},u.createElement(l.Tooltip,{title:"Tubular GitHub Repo"},u.createElement(l.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",color:"inherit"},u.createElement(s.default,null))),u.createElement(l.Tooltip,{title:this.state.openBasic?"Hide Code":"View Code"},u.createElement(l.IconButton,{onClick:this.handleClickBasic},u.createElement(o.default,null))),u.createElement(l.Tooltip,{title:"Edit in CodeSandbox"},u.createElement(l.IconButton,{component:"a",href:"https://codesandbox.io/s/1ox0q89rl7",target:"_blank",color:"inherit"},u.createElement(i.default,null)))),u.createElement(r.Collapse,{in:this.state.openBasic,timeout:"auto"},u.createElement(n.Paper,null,u.createElement(c.default,{language:"javascript",className:e.code},E.basicFeatures))),u.createElement(p.default,null))),u.createElement(n.Divider,null),u.createElement(n.Grid,{item:!0,xs:12,className:e.paper},u.createElement(n.Typography,{variant:"h5",gutterBottom:!0},"Give another presentation to your data."),u.createElement(n.Typography,{variant:"h6"},"You can provide to your data a styled view that allows you perform free-text searches and immediately reflect the changes over your data."),u.createElement(n.Grid,{item:!0,xs:12},u.createElement("div",{style:g},u.createElement(l.Tooltip,{title:"Tubular GitHub Repo"},u.createElement(l.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",color:"inherit"},u.createElement(s.default,null))),u.createElement(l.Tooltip,{title:this.state.openGridList?"Hide Code":"View Code"},u.createElement(l.IconButton,{onClick:this.handleClickGrid},u.createElement(o.default,null))),u.createElement(l.Tooltip,{title:"Edit in CodeSandbox"},u.createElement(l.IconButton,{component:"a",href:"https://codesandbox.io/s/8ywpr3ooq8",target:"_blank",color:"inherit"},u.createElement(i.default,null)))),u.createElement(r.Collapse,{in:this.state.openGridList,timeout:"auto"},u.createElement(n.Paper,null,u.createElement(c.default,{language:"javascript",className:e.code},E.gridList))),u.createElement(h.default,null))),u.createElement(n.Divider,null)))))},a}(u.Component));
-},{"@material-ui/core":"J2ra","@material-ui/icons/Code":"vRtn","@material-ui/icons/Create":"GAz6","react-highlight":"Xqwm","../components/Github":"qM49","react":"1n8/","highlight.js/styles/an-old-hope.css":"W7b0","../../srcdocs/components/BasicFeatures":"fiKH","../../srcdocs/components/SampleButtonsFeatures":"HqwG","../../srcdocs/components/SampleFeatures":"7E93","../../srcdocs/components/SampleGridList":"puvW","../utils/codeSnipetExamples":"Gxw2"}],"omJG":[function(require,module,exports) {
-"use strict";var e=this&&this.__extends||function(){var e=function(t,n){return(e=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var n in t)t.hasOwnProperty(n)&&(e[n]=t[n])})(t,n)};return function(t,n){function r(){this.constructor=t}e(t,n),t.prototype=null===n?Object.create(n):(r.prototype=n.prototype,new r)}}(),t=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)Object.hasOwnProperty.call(e,n)&&(t[n]=e[n]);return t.default=e,t},n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var r=t(require("react")),o=t(require("react-dom")),u=require("react-router"),a=require("react-router-dom"),c=n(require("../components/NavigationBar")),l=n(require("./Documentation/ColumnModel")),i=n(require("./Documentation/DataSource")),m=n(require("./Documentation/Getting-Started")),p=n(require("./Documentation/Home")),d=n(require("./Documentation/Props")),s=n(require("./Sample")),f=function(t){function n(){return null!==t&&t.apply(this,arguments)||this}return e(n,t),n.prototype.render=function(){return r.createElement(a.HashRouter,null,r.createElement(u.Switch,null,r.createElement("div",null,r.createElement(c.default,null),r.createElement(a.Route,{exact:!0,path:"/",component:p.default}),r.createElement(a.Route,{path:"/sample",component:s.default}),r.createElement(a.Redirect,{to:"/"}),r.createElement(a.HashRouter,null,r.createElement(u.Switch,null,r.createElement(a.Route,{path:"/documentation/props",component:d.default}),r.createElement(a.Route,{path:"/documentation/datasource",component:i.default}),r.createElement(a.Route,{path:"/documentation/getting-started",component:m.default}),r.createElement(a.Route,{path:"/documentation/columnmodel",component:l.default}))))))},n}(r.Component);o.render(r.createElement(f,null),document.getElementById("app"));
-},{"react":"1n8/","react-dom":"NKHc","react-router":"iOpc","react-router-dom":"obMO","../components/NavigationBar":"NH2g","./Documentation/ColumnModel":"q1fd","./Documentation/DataSource":"EGn3","./Documentation/Getting-Started":"VWQ5","./Documentation/Home":"9PWw","./Documentation/Props":"V9e7","./Sample":"rxup"}]},{},["omJG"], null)
+},{"@material-ui/core":"J2ra","date-fns":"cWQX","react":"1n8/","../utils/columns":"jRiX","../../src":"+fUd"}],"jOQw":[function(require,module,exports) {
+"use strict";var e=this&&this.__extends||function(){var e=function(t,a){return(e=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var a in t)t.hasOwnProperty(a)&&(e[a]=t[a])})(t,a)};return function(t,a){function n(){this.constructor=t}e(t,a),t.prototype=null===a?Object.create(a):(n.prototype=a.prototype,new n)}}(),t=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}},a=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var a in e)Object.hasOwnProperty.call(e,a)&&(t[a]=e[a]);return t.default=e,t};exports.__esModule=!0;var n=require("@material-ui/core"),r=require("@material-ui/core"),l=require("@material-ui/core"),o=t(require("@material-ui/icons/Code")),i=t(require("@material-ui/icons/Create")),c=t(require("react-highlight")),u=t(require("../../components/Github")),s=a(require("react"));require("highlight.js/styles/an-old-hope.css");var m=t(require("../../components/BasicFeatures")),p=t(require("../../components/SampleButtonsFeatures")),d=t(require("../../components/SampleFeatures")),h=t(require("../../components/SampleGridList")),E=require("../../utils/codeSnipetExamples"),f={code:{fontSize:15},container:{margin:"0",padding:30,width:"100%"},content:{marginTop:10},icon:{height:24,viewBox:"0 0 24 24",width:24},paper:{padding:20}},g={textAlign:"right"};exports.default=n.withStyles(f)(function(t){function a(){var e=null!==t&&t.apply(this,arguments)||this;return e.state={openBasic:!1,openGridList:!1,openSampleFeatures:!1,openSample:!1,showCode:!1,value:"one"},e.handleClickBasic=function(){e.setState({openBasic:!e.state.openBasic})},e.handleClickSample=function(){e.setState({openSample:!e.state.openSample})},e.handleClickSampleFeatures=function(){e.setState({openSampleFeatures:!e.state.openSampleFeatures})},e.handleClickGrid=function(){e.setState({openGridList:!e.state.openGridList})},e}return e(a,t),a.prototype.render=function(){var e=this.props.classes;return s.createElement("div",null,s.createElement(n.Grid,{container:!0,spacing:24,className:e.container},s.createElement(n.Grid,{item:!0,xs:12,className:e.paper},s.createElement(n.Paper,{className:e.paper},s.createElement(n.Typography,{variant:"h4"},"Samples"),s.createElement(n.Divider,null),s.createElement("br",null),s.createElement(n.Grid,{item:!0,xs:12,className:e.paper},s.createElement(n.Typography,{variant:"h5",gutterBottom:!0},"Very basic grid sample"),s.createElement(n.Typography,{variant:"h6"},"Basic grid with sorting, paging and filtering functionalities. Showing how different data types are displayed. To implement it you must declare each column and their own funcionalities."),s.createElement(n.Grid,{item:!0,xs:12},s.createElement("div",{style:g},s.createElement(l.Tooltip,{title:"Tubular GitHub Repo"},s.createElement(l.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",color:"inherit"},s.createElement(u.default,null))),s.createElement(l.Tooltip,{title:this.state.openBasic?"Hide Code":"View Code"},s.createElement(l.IconButton,{onClick:this.handleClickBasic},s.createElement(o.default,null))),s.createElement(l.Tooltip,{title:"Edit in CodeSandbox"},s.createElement(l.IconButton,{component:"a",href:"https://codesandbox.io/s/lp594wqr1q",target:"_blank",color:"inherit"},s.createElement(i.default,null)))),s.createElement(r.Collapse,{in:this.state.openBasic,timeout:"auto"},s.createElement(n.Paper,null,s.createElement(c.default,{language:"javascript",className:e.code},E.basicFeatures))),s.createElement(m.default,null))),s.createElement(n.Divider,null),s.createElement(n.Grid,{item:!0,xs:12,className:e.paper},s.createElement(n.Typography,{variant:"h5",gutterBottom:!0},"Full grid features"),s.createElement(n.Typography,{variant:"h6"},"The grid can be extended to include features like sorting, filtering, free-text searches, paging and change the number of rows per page. Also, you can easily print or export the current view or entire dataset to CSV using client-side only."),s.createElement(n.Grid,{item:!0,xs:12},s.createElement("div",{style:g},s.createElement(l.Tooltip,{title:"Tubular GitHub Repo"},s.createElement(l.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",color:"inherit"},s.createElement(u.default,null))),s.createElement(l.Tooltip,{title:this.state.openSample?"Hide Code":"View Code"},s.createElement(l.IconButton,{onClick:this.handleClickSample},s.createElement(o.default,null))),s.createElement(l.Tooltip,{title:"Edit in CodeSandbox"},s.createElement(l.IconButton,{component:"a",href:"https://codesandbox.io/s/434rlkr2y9",target:"_blank",color:"inherit"},s.createElement(i.default,null)))),s.createElement(r.Collapse,{in:this.state.openSample,timeout:"auto"},s.createElement(n.Paper,null,s.createElement(c.default,{language:"javascript",className:e.code},E.simpleFeatures))),s.createElement(d.default,null))),s.createElement(n.Divider,null),s.createElement(n.Grid,{item:!0,xs:12,className:e.paper},s.createElement(n.Typography,{variant:"h5",gutterBottom:!0},"How to include functionality buttons"),s.createElement(n.Typography,{variant:"h6"},"You can add functionalities to the DataGrid including extra buttons that can perform an action according to your requirements. You just need to include the IconButton Component from @material-ui and define the icon or button that you need between DataGrid tags and specify the action to perform."),s.createElement(n.Grid,{item:!0,xs:12},s.createElement("div",{style:g},s.createElement(l.Tooltip,{title:"Tubular GitHub Repo"},s.createElement(l.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",color:"inherit"},s.createElement(u.default,null))),s.createElement(l.Tooltip,{title:this.state.openSampleFeatures?"Hide Code":"View Code"},s.createElement(l.IconButton,{onClick:this.handleClickSampleFeatures},s.createElement(o.default,null))),s.createElement(l.Tooltip,{title:"Edit in CodeSandbox"},s.createElement(l.IconButton,{component:"a",href:"https://codesandbox.io/s/1ox0q89rl7",target:"_blank",color:"inherit"},s.createElement(i.default,null)))),s.createElement(r.Collapse,{in:this.state.openSampleFeatures,timeout:"auto"},s.createElement(n.Paper,null,s.createElement(c.default,{language:"javascript",className:e.code},E.buttonFeatures))),s.createElement(p.default,null))),s.createElement(n.Divider,null),s.createElement(n.Grid,{item:!0,xs:12,className:e.paper},s.createElement(n.Typography,{variant:"h5",gutterBottom:!0},"Give another presentation to your data."),s.createElement(n.Typography,{variant:"h6"},"You can provide to your data a styled view that allows you perform free-text searches and immediately reflect the changes over your data."),s.createElement(n.Grid,{item:!0,xs:12},s.createElement("div",{style:g},s.createElement(l.Tooltip,{title:"Tubular GitHub Repo"},s.createElement(l.IconButton,{component:"a",href:"https://github.com/unosquare/tubular-react",target:"_blank",color:"inherit"},s.createElement(u.default,null))),s.createElement(l.Tooltip,{title:this.state.openGridList?"Hide Code":"View Code"},s.createElement(l.IconButton,{onClick:this.handleClickGrid},s.createElement(o.default,null))),s.createElement(l.Tooltip,{title:"Edit in CodeSandbox"},s.createElement(l.IconButton,{component:"a",href:"https://codesandbox.io/s/8ywpr3ooq8",target:"_blank",color:"inherit"},s.createElement(i.default,null)))),s.createElement(r.Collapse,{in:this.state.openGridList,timeout:"auto"},s.createElement(n.Paper,null,s.createElement(c.default,{language:"javascript",className:e.code},E.gridList))),s.createElement(h.default,null))),s.createElement(n.Divider,null)))))},a}(s.Component));
+},{"@material-ui/core":"J2ra","@material-ui/icons/Code":"vRtn","@material-ui/icons/Create":"GAz6","react-highlight":"Xqwm","../../components/Github":"qM49","react":"1n8/","highlight.js/styles/an-old-hope.css":"W7b0","../../components/BasicFeatures":"fiKH","../../components/SampleButtonsFeatures":"HqwG","../../components/SampleFeatures":"7E93","../../components/SampleGridList":"puvW","../../utils/codeSnipetExamples":"Gxw2"}],"WKKG":[function(require,module,exports) {
+"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var l={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(l[t]=e[t]);return l.default=e,l};exports.__esModule=!0;var l=require("@material-ui/core"),t=require("@material-ui/core/styles"),a=e(require("react")),n=function(e){var l=e.spacing;return t.createStyles({code:{background:"#F8F8FF",fontSize:14},root:{marginTop:3*l.unit,overflowX:"auto",width:"100%"}})},r=function(e){var t=e.classes;return a.createElement(l.Paper,{className:t.root},a.createElement(l.Table,null,a.createElement(l.TableHead,null,a.createElement(l.TableRow,null,a.createElement(l.TableCell,null,"Name"),a.createElement(l.TableCell,null,"Type"),a.createElement(l.TableCell,null,"Default value"),a.createElement(l.TableCell,null,"Description"))),a.createElement(l.TableBody,null,a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"gridName")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"string")),a.createElement(l.TableCell,null,"Grid"),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"toolbarOptions")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"ToolbarOptions")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"new ToolBarOptions();")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional."),"It should be an instance of ToolbarOptions. This encapsulates useful options.",a.createElement("em",null,"see ToolBarOptions props"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bodyRenderer")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"function")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional."),"It takes a function with 3 parameters",a.createElement("code",{className:t.code},"(row: any, index: number, columns: ColumnModel[])")," to map all rows.",a.createElement("em",null,"see sample Full grid features"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"footerRenderer")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"function")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional."),"It takes an aggregate function with 1 parameter",a.createElement("code",{className:t.code},"(aggregate: any)")," to show in a foot row the results from the aggregate function.",a.createElement("em",null,"see sample Full grid features"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"onRowClick")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"function")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional."),"Use this event handler to receive the clicked row.")))))};exports.default=t.withStyles(n)(r);
+},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/"}],"AkVR":[function(require,module,exports) {
+"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var l={};if(null!=e)for(var a in e)Object.hasOwnProperty.call(e,a)&&(l[a]=e[a]);return l.default=e,l};exports.__esModule=!0;var l=require("@material-ui/core"),a=e(require("react")),t=function(e){return l.createStyles({code:{background:"#F8F8FF",fontSize:14},root:{marginTop:3*e.spacing.unit,overflowX:"auto",width:"100%"}})},c=function(e){var t=e.classes;return a.createElement(l.Table,null,a.createElement(l.TableHead,null,a.createElement(l.TableRow,null,a.createElement(l.TableCell,null,"Name"),a.createElement(l.TableCell,null,"Type"),a.createElement(l.TableCell,null,"Default"),a.createElement(l.TableCell,null,"Description"),a.createElement(l.TableCell,null,"Options"))),a.createElement(l.TableBody,null,a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Name")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"string")),a.createElement(l.TableCell,null,"NONE"),a.createElement(l.TableCell,null,"This is required and represents a property of the entity which we are binding"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Aggregate")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"AggregateFunctions")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"NONE")),a.createElement(l.TableCell,null,"The aggregation function that will be applied to this column"),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"NONE"),",",a.createElement("code",{className:t.code},"SUM"),",",a.createElement("code",{className:t.code},"AVERAGE"),",",a.createElement("br",null),a.createElement("code",{className:t.code},"COUNT"),",",a.createElement("code",{className:t.code},"DISTINCT_COUNT"),",",a.createElement("br",null),a.createElement("code",{className:t.code},"MAX"),",",a.createElement("code",{className:t.code},"MIN"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"DataType")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"ColumnDataType")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"string")),a.createElement(l.TableCell,null,"The column type"),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"STRING"),",",a.createElement("code",{className:t.code},"NUMERIC"),",",a.createElement("code",{className:t.code},"BOOLEAN"),",",a.createElement("br",null),a.createElement("code",{className:t.code},"DATE"),",",a.createElement("code",{className:t.code},"DATE_TIME"),",",a.createElement("code",{className:t.code},"DATE_TIME_UTC"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Filterable")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"false")),a.createElement(l.TableCell,null,"Enables Filterable"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"IsKey")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"false")),a.createElement(l.TableCell,null,"Defines if a column is an identifier or not"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Label")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"string")),a.createElement(l.TableCell,null,"The name of the column"),a.createElement(l.TableCell,null,"Column label that will be shown"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Searchable")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,"Indicates that a column can be used to search upon"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"SortDirection")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"ColumnSortDirection")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"NONE")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"NONE"),",",a.createElement("code",{className:t.code},"ASCENDING"),",",a.createElement("code",{className:t.code},"DESCENDING"))),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"SortOrder")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"number")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"-1")),a.createElement(l.TableCell,null,"-"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Sortable")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"false")),a.createElement(l.TableCell,null,"Determines if a column can be sorted"),a.createElement(l.TableCell,null,"-")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"Visible")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bool")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,"Specifies if a column should be shown"),a.createElement(l.TableCell,null,"-"))))};exports.default=l.withStyles(t)(c);
+},{"@material-ui/core":"J2ra","react":"1n8/"}],"AtZZ":[function(require,module,exports) {
+"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var l={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(l[t]=e[t]);return l.default=e,l};exports.__esModule=!0;var l=require("@material-ui/core"),t=require("@material-ui/core/styles"),a=e(require("react")),n=function(e){return t.createStyles({code:{background:"#F8F8FF",fontSize:14},root:{marginTop:3*e.spacing.unit,overflowX:"auto",width:"100%"}})},r=function(e){var t=e.classes;return a.createElement(l.Paper,{className:t.root},a.createElement(l.Table,null,a.createElement(l.TableHead,null,a.createElement(l.TableRow,null,a.createElement(l.TableCell,null,"Name"),a.createElement(l.TableCell,null,"Type"),a.createElement(l.TableCell,null,"Default value"),a.createElement(l.TableCell,null,"Description"))),a.createElement(l.TableBody,null,a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"advancePagination")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows advanced pagination with numbers of the page to navigate between pages, also display the navigation buttons to the last or first page. Otherwise, show a simple pagination with two directions arrows for advance or go back a page.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"bottomPager")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows pagination bar in the bottom of the grid.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"exportButton")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows the export button.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"printButton")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows the print button.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"searchText")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows the search text input.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"topPager")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"boolean")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"true")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows pagination bar in the top of the grid.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"rowsPerPageOptions")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"array")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"[10, 20, 50, 100]")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"It shows a combo with a set of values that represent the number of rows per page.")),a.createElement(l.TableRow,{hover:!0},a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"rowsPerPage")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"number")),a.createElement(l.TableCell,null,a.createElement("code",{className:t.code},"10")),a.createElement(l.TableCell,null,a.createElement("strong",null,"Optional. "),"Initial valor that represent the number of rows per page. It should be a number that is inside the ",a.createElement("em",null,"rowsPerPageOptions")," array.")))))};exports.default=t.withStyles(n)(r);
+},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/"}],"xpDp":[function(require,module,exports) {
+"use strict";var e=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var a={};if(null!=e)for(var t in e)Object.hasOwnProperty.call(e,t)&&(a[t]=e[t]);return a.default=e,a},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var t=require("@material-ui/core"),r=require("@material-ui/core/styles"),o=e(require("react")),n=a(require("react-highlight")),l=a(require("../../components/DataGridProps")),c=a(require("../../components/ColumnModelTable")),s=a(require("../../components/ToolBarOptionsProps")),i={codeTag:{background:"#F8F8FF",padding:"6px"},container:{margin:"0",padding:30,width:"100%"},paper:{padding:10}},m="... \nconst MyComponent = () => {\n        return (\n            <div className='root'>\n                <DataGrid\n                    gridName='Tubular-React'\n                />\n            </div>\n        );\n};\n\n\n/*\n Use the component withRemoteDataSource to wrap your component\n and columns definition among the data obtained from the URL.\n\n withRemoteDataSource will set an initial context for your grid.\n*/\nexport default withRemoteDataSource(BasicFeatures, columns, 'https://tubular.azurewebsites.net/api/orders/paged');\n",d="... \nconst MyComponent = () => {\n        return (\n            <div className='root'>\n                <DataGrid\n                    gridName='Tubular-React'\n                />\n            </div>\n        );\n};\n\n\nexport default withRemoteDataSource(BasicFeatures, columns, localData)",u="https://github.com/unosquare/tubular-react/blob/master/sample/src/data/localData.ts",p=function(e){var a=e.classes;return o.createElement(t.Grid,{container:!0,spacing:24,className:a.container},o.createElement(t.Grid,{item:!0,xs:12},o.createElement(t.Paper,{className:a.paper,style:{overflowX:"auto"}},o.createElement(t.Typography,{variant:"h4",paragraph:!0}," Column Model "),o.createElement(t.Divider,null),o.createElement(c.default,null))),o.createElement(t.Grid,{item:!0,xs:12},o.createElement(t.Paper,{className:a.paper},o.createElement(t.Typography,{variant:"h4",paragraph:!0},"DataGrid props"),o.createElement(t.Divider,null),o.createElement("br",null),o.createElement(t.Typography,{variant:"subtitle1"},"It's important to use ",o.createElement("code",{className:a.codeTag},"withRemoteDataSource")," or",o.createElement("code",{className:a.codeTag},"withLocalDataSource")," according to the case to fill the ",o.createElement("code",{className:a.codeTag},"<DataGrid />")," component with the data.",o.createElement("br",null),"These are all the available props (and their default values) for the",o.createElement("code",{className:a.codeTag},"<DataGrid />")," component."),o.createElement(l.default,null),o.createElement("br",null),o.createElement(t.Typography,{variant:"subtitle1"},o.createElement("i",null,"If you don't define some of the optional props described above, these will not be shown. In the case of "),o.createElement("code",{className:a.codeTag},"bodyRenderer"),",",o.createElement("i",null,"the grid will display its default body; if the",o.createElement("code",{className:a.codeTag},"footerRenderer"),"is not defined, the footer will not be displayed.")))),o.createElement(t.Grid,{item:!0,xs:12},o.createElement(t.Paper,{className:a.paper},o.createElement(t.Typography,{variant:"h4",paragraph:!0},"Data Source"),o.createElement(t.Divider,null),o.createElement("br",null),o.createElement(t.Typography,{variant:"subtitle1"},o.createElement("code",{className:a.codeTag},"<DataGrid />")," requires a wrapper component to retrieve data. Depends on if we use a local data source or a remote data source, these wrappers are:",o.createElement("code",{className:a.codeTag},"withRemoteDataSource"),"component or the",o.createElement("code",{className:a.codeTag},"withLocalDataSource"),"component, which deals with data retrieval on the",o.createElement("code",{className:a.codeTag},"DataGrid")," component.",o.createElement("br",null),o.createElement("br",null),o.createElement(t.Typography,{variant:"h5",gutterBottom:!0},o.createElement("code",{className:a.codeTag},"withRemoteDataSource")),o.createElement("code",{className:a.codeTag},"withRemoteDataSource"),"needs our ",o.createElement("code",{className:a.codeTag},"DataGrid"),"component, a ",o.createElement("code",{className:a.codeTag},"ColumnModel"),"array and a ",o.createElement("code",{className:a.codeTag},"URL")," which represent a service.",o.createElement("br",null),o.createElement(n.default,{language:"javascript",className:"an-old-hope"},m),o.createElement("br",null),o.createElement(t.Typography,{variant:"h5",gutterBottom:!0},o.createElement("code",{className:a.codeTag},"LocalDataSource")),o.createElement("code",{className:a.codeTag},"withLocalDataSource"),"needs our ",o.createElement("code",{className:a.codeTag},"DataGrid"),"component, a ",o.createElement("code",{className:a.codeTag},"ColumnModel"),"array and an ",o.createElement("code",{className:a.codeTag},"Array"),"of objects.",o.createElement("br",null),o.createElement(n.default,{language:"javascript",className:"an-old-hope"},d),"See this",o.createElement(t.Button,{color:"primary",href:u,target:"_blank"},"example"),"of how to define the array of objects."))),o.createElement(t.Grid,{item:!0,xs:12},o.createElement(t.Paper,{className:a.paper},o.createElement(t.Typography,{variant:"h4",paragraph:!0},"Toolbar options"),o.createElement(t.Divider,null),o.createElement("br",null),o.createElement(t.Typography,{variant:"subtitle1"},"If you need to personalize the grid adding, removing or modifying  features, ",o.createElement("code",{className:a.codeTag},"ToolBarOptions Class"),"provides several options."),o.createElement(s.default,null),o.createElement("br",null),o.createElement(t.Typography,{variant:"subtitle1"},o.createElement("i",null,"If you don't define some of the optional props described above, these will set with the default values.")))))};exports.default=r.withStyles(i)(p);
+},{"@material-ui/core":"J2ra","@material-ui/core/styles":"+Ptv","react":"1n8/","react-highlight":"Xqwm","../../components/DataGridProps":"WKKG","../../components/ColumnModelTable":"AkVR","../../components/ToolBarOptionsProps":"AtZZ"}],"omJG":[function(require,module,exports) {
+"use strict";var e=this&&this.__extends||function(){var e=function(t,r){return(e=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var r in t)t.hasOwnProperty(r)&&(e[r]=t[r])})(t,r)};return function(t,r){function n(){this.constructor=t}e(t,r),t.prototype=null===r?Object.create(r):(n.prototype=r.prototype,new n)}}(),t=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var r in e)Object.hasOwnProperty.call(e,r)&&(t[r]=e[r]);return t.default=e,t},r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};exports.__esModule=!0;var n=t(require("react")),o=t(require("react-dom")),u=require("react-router"),a=require("react-router-dom"),c=r(require("../components/NavigationBar")),i=r(require("./Documentation/Home")),l=r(require("./Documentation/Sample")),p=r(require("./Documentation/ComponentAPI")),m=function(t){function r(){return null!==t&&t.apply(this,arguments)||this}return e(r,t),r.prototype.render=function(){return n.createElement(a.HashRouter,null,n.createElement(u.Switch,null,n.createElement("div",null,n.createElement(c.default,null),n.createElement(a.Route,{exact:!0,path:"/",component:i.default}),n.createElement(a.Route,{path:"/sample",component:l.default}),n.createElement(a.Route,{path:"/componentAPI/",component:p.default}))))},r}(n.Component);o.render(n.createElement(m,null),document.getElementById("app"));
+},{"react":"1n8/","react-dom":"NKHc","react-router":"iOpc","react-router-dom":"obMO","../components/NavigationBar":"NH2g","./Documentation/Home":"9PWw","./Documentation/Sample":"jOQw","./Documentation/ComponentAPI":"xpDp"}]},{},["omJG"], null)
