+++ srcdocs/index.html
@@ -24,4 +24,8 @@
     <div id="app"></div>
     <script src="./pages/index.tsx" type="text/javascript"></script>
   </body>
+
+  <script>
+    (function (i, s, o, g, r, a, m) { i['GoogleAnalyticsObject'] = r; i[r] = i[r] || function () { (i[r].q = i[r].q || []).push(arguments) }, i[r].l = 1 * new Date(); a = s.createElement(o), m = s.getElementsByTagName(o)[0]; a.async = 1; a.src = g; m.parentNode.insertBefore(a, m) })(window, document, 'script', '//www.google-analytics.com/analytics.js', 'ga'); ga('create', 'UA-8535255-1', 'unosquare.com'); ga('send', 'pageview');
+</script>
 </html>
