--- srcdocs/pages/Documentation/DataSource.tsx
@@ -1,103 +1,0 @@
-import {
-  Button,
-  Divider,
-  Grid,
-  Hidden,
-  Paper,
-  Typography,
-  WithStyles,
-  withStyles,
-} from '@material-ui/core';
-import 'highlight.js/styles/an-old-hope.css';
-import * as React from 'react';
-import Highlight from 'react-highlight';
-import DocumentationList from '../../components/DocumentationList';
-
-
-const styles = {  
-  codeTag: {
-    background: '#F8F8FF',
-    padding: '6px',
-  },
-  container: {
-    margin: '0',
-    padding: 30,
-    width: '100%',
-  },
-  paper: {
-    padding: 10,
-  },
-};
-
-// tslint:disable-next-line:max-line-length
-const remoteDataSource =
-  "... \nconst MyComponent = () => {\n        return (\n            <div className='root'>\n                <DataGrid\n                    gridName='Tubular-React'\n                />\n            </div>\n        );\n};\n\n\n/*\n Use the component withRemoteDataSource to wrap your component\n and columns definition among the data obtained from the URL.\n\n withRemoteDataSource will set an initial context for your grid.\n*/\nexport default withRemoteDataSource(BasicFeatures, columns, 'https://tubular.azurewebsites.net/api/orders/paged');\n";
-// tslint:disable-next-line:max-line-length
-const localDataSource =
-  "... \nconst MyComponent = () => {\n        return (\n            <div className='root'>\n                <DataGrid\n                    gridName='Tubular-React'\n                />\n            </div>\n        );\n};\n\n\nexport default withRemoteDataSource(BasicFeatures, columns, localData)";
-// tslint:disable-next-line:max-line-length
-const localDataLink =
-  'https://github.com/unosquare/tubular-react/blob/master/sample/src/data/localData.ts';
-
-const DataSource = ({ classes }: any) => (
-  <Grid container={true} spacing={24} className={classes.container}>
-    <Hidden smDown={true}>
-      <Grid item={true} xs={3}>
-        <DocumentationList />
-      </Grid>
-    </Hidden>
-    <Grid item={true} xs={12} md={9}>
-      <Paper className={classes.paper}>
-        <Typography variant='h4' paragraph={true}>
-          Data Source
-        </Typography>
-        <Divider />
-        <br />
-        <Typography variant='subtitle1'>
-          <code className={classes.codeTag}>{'<DataGrid />'}</code> requires a
-          wrapper component to retrieve data. Depends on if we use a local
-          data source or a remote data source, these wrappers are:
-              <code className={classes.codeTag}>withRemoteDataSource</code>
-          component or the
-          <code className={classes.codeTag}>withLocalDataSource</code>
-          component, which deals with data retrieval on the
-              <code className={classes.codeTag}>DataGrid</code> component.
-              <br />
-          <br />
-          <Typography variant='h5' gutterBottom={true}>
-            <code className={classes.codeTag}>withRemoteDataSource</code>
-          </Typography>
-          <code className={classes.codeTag}>withRemoteDataSource</code>
-          needs our <code className={classes.codeTag}>DataGrid</code>
-          component, a <code className={classes.codeTag}>ColumnModel</code>
-          array and a <code className={classes.codeTag}>URL</code> which
-          represent a service.
-              <br />
-          <Highlight language='javascript' className={'an-old-hope'}>
-            {remoteDataSource}
-          </Highlight>
-          <br />
-          <Typography variant='h5' gutterBottom={true}>
-            <code className={classes.codeTag}>LocalDataSource</code>
-          </Typography>
-          <code className={classes.codeTag}>withLocalDataSource</code>
-          needs our <code className={classes.codeTag}>DataGrid</code>
-          component, a <code className={classes.codeTag}>ColumnModel</code>
-          array and an <code className={classes.codeTag}>Array</code>of
-          objects.
-              <br />
-          <Highlight language='javascript' className={'an-old-hope'}>
-            {localDataSource}
-          </Highlight>
-          See this
-              <Button color='primary' href={localDataLink} target='_blank'>
-            example
-              </Button>
-          of how to define the array of objects.
-            </Typography>
-      </Paper>
-    </Grid>
-  </Grid>
-);
-
-export default withStyles(styles)(DataSource);
