+++ srcdocs/pages/Index.tsx
@@ -1,33 +1,22 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { Switch } from 'react-router';
-import { HashRouter, Redirect, Route } from 'react-router-dom';
+import { HashRouter, Route } from 'react-router-dom';
 import NavigationBar from '../components/NavigationBar';
-import ColumnModel from './Documentation/ColumnModel';
-import DataSource from './Documentation/DataSource';
-import GettingStarted from './Documentation/Getting-Started';
 import Home from './Documentation/Home';
-import Props from './Documentation/Props';
-import Sample from './Sample';
+import Sample from './Documentation/Sample';
+import ComponentAPI from './Documentation/ComponentAPI';
 
 class App extends React.Component {
-   public render() {
+  public render() {
     return (
       <HashRouter>
-         <Switch>
+        <Switch>
           <div>
-            <NavigationBar/>
-            <Route exact={true} path='/' component={Home}/>
-            <Route  path='/sample' component={Sample}/>
-            <Redirect to='/' />
-            <HashRouter>
-              <Switch>
-                <Route path='/documentation/props' component={Props}/>
-                <Route  path='/documentation/datasource' component={DataSource}/>
-                <Route  path='/documentation/getting-started' component={GettingStarted}/>
-                <Route  path='/documentation/columnmodel' component={ColumnModel}/>
-              </Switch>
-            </HashRouter>
+            <NavigationBar />
+            <Route exact={true} path='/' component={Home} />
+            <Route path='/sample' component={Sample} />
+            <Route path='/componentAPI/' component={ComponentAPI} />
           </div>
         </Switch>
       </HashRouter>
@@ -35,4 +24,4 @@ class App extends React.Component {
   }
 }
 
-ReactDOM.render(<App/>, document.getElementById('app'));
+ReactDOM.render(<App />, document.getElementById('app'));
